refactor(platform): remove chat run work folding switch
