+++ turbo/apps/platform/src/i18n/locales/de-DE/common.json
@@ -2150,26 +2150,17 @@
     "promptDocumentTitle": "Prompt",
     "queue": {
       "aboutAutomationEvent": "Über dieses Automatisierungsereignis",
-      "aboutGoal": "Über dieses Ziel",
       "aboutQueuedMessage": "Über diese Nachricht in der Warteschlange",
-      "activeGoal": "Aktives Ziel",
       "automationEvent": "Automatisierungsereignis",
       "automationEventDescription": "Wartet hinter Nachrichten in der Warteschlange und wird ausgeführt, sobald die aktuelle Ausführung abgeschlossen ist.",
-      "cancelGoal": "Ziel abbrechen",
       "cancellationRecoveryPending": "Die abgebrochene Ausführung wird abgeschlossen, bevor die Arbeit in der Warteschlange fortgesetzt wird.",
       "event_one": "{{count}} Ereignis",
       "event_other": "{{count}} Ereignisse",
       "goal": "Ziel",
-      "goalDescription": "Wird ausgeführt, nachdem die Warteschlange geleert wurde, und läuft weiter, bis Sie es abbrechen.",
-      "goalLoadFailed": "Dieses Ziel konnte nicht geladen werden",
-      "goalRetry": "Schließen Sie das Dialogfeld und öffnen Sie es erneut, um es erneut zu versuchen.",
-      "goalUnavailable": "Dieses Ziel ist nicht mehr verfügbar.",
       "itemsWaiting": "{{items}} wartet",
       "itemsWaitingTogether": "{{messages}} und {{events}} warten",
-      "loadingGoal": "Ziel wird geladen...",
       "message_one": "{{count}} Nachricht",
       "message_other": "{{count}} Nachrichten",
-      "openGoalDetails": "Zieldetails öffnen",
       "pendingAutomationEvent": "Ausstehendes Automatisierungsereignis",
       "queuedMessage": "Nachricht in der Warteschlange",
       "queuedMessageDescription": "Wartet in der Warteschlange und wird gesendet, sobald die aktuelle Ausführung beendet ist.",
@@ -2181,8 +2172,6 @@
       "title": "Composer-Entwurf ersetzen?"
     },
     "run": {
-      "automatedRun": "Automatisierte Ausführung",
-      "collapseGroupedHistory": "Gruppierten Ausführungsverlauf einklappen",
       "collapseWorkHistory": "Arbeitsverlauf einklappen",
       "creditUsage": "Credit-Nutzung",
       "done": {
@@ -2197,28 +2186,19 @@
         "wrappedUp": "Abgeschlossen um {{time}}"
       },
       "duration": {
-        "hour_one": "{{count}} Stunde",
-        "hour_other": "{{count}} Stunden",
         "hoursShort_one": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} Min.",
-        "minute_other": "{{count}} Min.",
         "minutesShort_one": "{{count}}m",
         "minutesShort_other": "{{count}}m",
         "secondsShort_one": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{duration}} für {{label}}",
-      "expandGroupedHistory": "Gruppierten Ausführungsverlauf ausklappen",
       "expandWorkHistory": "Arbeitsverlauf ausklappen",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Fast-Modus deaktiviert",
       "fastModeOn": "Fast-Modus aktiviert",
       "fastModeWillBeOff": "Fast-Modus wird deaktiviert",
       "fastModeWillBeOn": "Fast-Modus wird aktiviert",
-      "goalFor": "Ziel für {{label}}",
-      "groupedRunsFor_one": "{{count}} Ausführung für {{source}}",
-      "groupedRunsFor_other": "{{count}} Ausführungen für {{source}}",
       "justNow": "soeben",
       "keepGoing": "Fortsetzen",
       "keepGoingAt": "Fortsetzen · {{timestamp}}",
@@ -2243,7 +2223,6 @@
       "viewActivityLogs": "Aktivitätsprotokolle anzeigen",
       "viewLogs": "Ausführungsprotokolle anzeigen",
       "waitingIn": "Warten",
-      "worked": "Ausgeführt",
       "workedFor": "Ausgeführt in {{duration}}",
       "workingFor": "Seit {{duration}} in Arbeit"
     },
