+++ turbo/apps/platform/src/i18n/locales/en-US/common.json
@@ -2150,26 +2150,17 @@
     "promptDocumentTitle": "Prompt",
     "queue": {
       "aboutAutomationEvent": "About this automation event",
-      "aboutGoal": "About this goal",
       "aboutQueuedMessage": "About this queued message",
-      "activeGoal": "Active goal",
       "automationEvent": "Automation event",
       "automationEventDescription": "Waits behind queued messages and runs once the current run finishes.",
-      "cancelGoal": "Cancel goal",
       "cancellationRecoveryPending": "Finalizing the cancelled run before queued work continues.",
       "event_one": "{{count}} event",
       "event_other": "{{count}} events",
       "goal": "Goal",
-      "goalDescription": "Runs after the queue drains and keeps running until you cancel it.",
-      "goalLoadFailed": "Couldn't load this goal",
-      "goalRetry": "Close the dialog and open it again to retry.",
-      "goalUnavailable": "This goal is no longer available.",
       "itemsWaiting": "{{items}} waiting",
       "itemsWaitingTogether": "{{messages}} and {{events}} waiting",
-      "loadingGoal": "Loading goal...",
       "message_one": "{{count}} message",
       "message_other": "{{count}} messages",
-      "openGoalDetails": "Open goal details",
       "pendingAutomationEvent": "Pending automation event",
       "queuedMessage": "Queued message",
       "queuedMessageDescription": "Waits in line and sends once the current run finishes.",
@@ -2181,8 +2172,6 @@
       "title": "Replace composer draft?"
     },
     "run": {
-      "automatedRun": "Automated run",
-      "collapseGroupedHistory": "Collapse grouped run history",
       "collapseWorkHistory": "Collapse work history",
       "creditUsage": "Credit usage",
       "done": {
@@ -2197,28 +2186,19 @@
         "wrappedUp": "Wrapped up at {{time}}"
       },
       "duration": {
-        "hour_one": "{{count}} hour",
-        "hour_other": "{{count}} hours",
         "hoursShort_one": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} min",
-        "minute_other": "{{count}} mins",
         "minutesShort_one": "{{count}}m",
         "minutesShort_other": "{{count}}m",
         "secondsShort_one": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{duration}} for {{label}}",
-      "expandGroupedHistory": "Expand grouped run history",
       "expandWorkHistory": "Expand work history",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Fast mode off",
       "fastModeOn": "Fast mode on",
       "fastModeWillBeOff": "Fast mode will be off",
       "fastModeWillBeOn": "Fast mode will be on",
-      "goalFor": "Goal for {{label}}",
-      "groupedRunsFor_one": "{{count}} run for {{source}}",
-      "groupedRunsFor_other": "{{count}} runs for {{source}}",
       "justNow": "just now",
       "keepGoing": "Keep going",
       "keepGoingAt": "Keep going · {{timestamp}}",
@@ -2243,7 +2223,6 @@
       "viewActivityLogs": "View activity logs",
       "viewLogs": "View run logs",
       "waitingIn": "Waiting in",
-      "worked": "Worked",
       "workedFor": "Worked for {{duration}}",
       "workingFor": "Working for {{duration}}"
     },
