+++ turbo/apps/platform/src/i18n/locales/es-ES/common.json
@@ -2184,28 +2184,19 @@
     "promptDocumentTitle": "Prompt",
     "queue": {
       "aboutAutomationEvent": "Acerca de este evento de automatización",
-      "aboutGoal": "Sobre este objetivo",
       "aboutQueuedMessage": "Acerca de este mensaje en cola",
-      "activeGoal": "Objetivo activo",
       "automationEvent": "Evento de automatización",
       "automationEventDescription": "Espera detrás de los mensajes en cola y se ejecuta cuando termina la ejecución actual.",
-      "cancelGoal": "Cancelar objetivo",
       "cancellationRecoveryPending": "Finalizando la ejecución cancelada antes de continuar con el trabajo en cola.",
       "event_one": "{{count}} evento",
       "event_many": "{{count}} eventos",
       "event_other": "{{count}} eventos",
       "goal": "Objetivo",
-      "goalDescription": "Se ejecuta después de que la cola se vacía y sigue funcionando hasta que la cancelas.",
-      "goalLoadFailed": "No se pudo cargar este objetivo",
-      "goalRetry": "Cierra el diálogo y ábrelo de nuevo para intentarlo de nuevo.",
-      "goalUnavailable": "Este objetivo ya no está disponible.",
       "itemsWaiting": "{{items}} esperando",
       "itemsWaitingTogether": "{{messages}} y {{events}} esperando",
-      "loadingGoal": "Cargando objetivo...",
       "message_one": "{{count}} mensaje",
       "message_many": "{{count}} mensajes",
       "message_other": "{{count}} mensajes",
-      "openGoalDetails": "Abrir detalles del objetivo",
       "pendingAutomationEvent": "Evento de automatización pendiente",
       "queuedMessage": "Mensaje en cola",
       "queuedMessageDescription": "Espera en la cola y se envía cuando termina la ejecución actual.",
@@ -2217,8 +2208,6 @@
       "title": "¿Reemplazar el borrador del editor?"
     },
     "run": {
-      "automatedRun": "Ejecución automatizada",
-      "collapseGroupedHistory": "Contraer el historial de ejecuciones agrupadas",
       "collapseWorkHistory": "Contraer el historial de trabajo",
       "creditUsage": "Consumo de créditos",
       "done": {
@@ -2233,15 +2222,9 @@
         "wrappedUp": "Terminado en {{time}}"
       },
       "duration": {
-        "hour_one": "{{count}} hora",
-        "hour_many": "{{count}} horas",
-        "hour_other": "{{count}} horas",
         "hoursShort_one": "{{count}}h",
         "hoursShort_many": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} min",
-        "minute_many": "{{count}} minutos",
-        "minute_other": "{{count}} minutos",
         "minutesShort_one": "{{count}}m",
         "minutesShort_many": "{{count}}m",
         "minutesShort_other": "{{count}}m",
@@ -2249,18 +2232,12 @@
         "secondsShort_many": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{duration}} para {{label}}",
-      "expandGroupedHistory": "Expandir el historial de ejecuciones agrupadas",
       "expandWorkHistory": "Expandir el historial de trabajo",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Modo rápido desactivado",
       "fastModeOn": "Modo rápido activado",
       "fastModeWillBeOff": "El modo rápido se desactivará",
       "fastModeWillBeOn": "El modo rápido se activará",
-      "goalFor": "Objetivo para {{label}}",
-      "groupedRunsFor_one": "{{count}} ejecución desde {{source}}",
-      "groupedRunsFor_many": "{{count}} ejecuciones desde {{source}}",
-      "groupedRunsFor_other": "{{count}} ejecuciones desde {{source}}",
       "justNow": "Justo ahora",
       "keepGoing": "Sigue",
       "keepGoingAt": "Sigue adelante · {{timestamp}}",
@@ -2286,7 +2263,6 @@
       "viewActivityLogs": "Ver registros de actividad",
       "viewLogs": "Ver registros de ejecución",
       "waitingIn": "Esperando",
-      "worked": "Trabajo realizado",
       "workedFor": "Trabajo realizado durante {{duration}}",
       "workingFor": "Trabajando durante {{duration}}"
     },
