+++ turbo/apps/platform/src/i18n/locales/fr-FR/common.json
@@ -2184,28 +2184,19 @@
     "promptDocumentTitle": "Prompt",
     "queue": {
       "aboutAutomationEvent": "À propos de cet événement d'automatisation",
-      "aboutGoal": "À propos de cet objectif",
       "aboutQueuedMessage": "À propos de ce message en file d'attente",
-      "activeGoal": "Objectif actif",
       "automationEvent": "Événement d'automatisation",
       "automationEventDescription": "Attend derrière les messages en file d'attente et s'exécute une fois que l'exécution en cours est terminée.",
-      "cancelGoal": "Annuler l'objectif",
       "cancellationRecoveryPending": "Finalisation de l'exécution annulée avant la reprise du travail en file d'attente.",
       "event_one": "{{count}} événement",
       "event_many": "{{count}} événements",
       "event_other": "{{count}} événements",
       "goal": "Objectif",
-      "goalDescription": "S'exécute après que la file d'attente est vidée et continue à s'exécuter jusqu'à ce que vous l'annuliez.",
-      "goalLoadFailed": "Impossible de charger cet objectif",
-      "goalRetry": "Fermez la boîte de dialogue et rouvrez-la pour réessayer.",
-      "goalUnavailable": "Cet objectif n'est plus disponible.",
       "itemsWaiting": "{{items}} en attente",
       "itemsWaitingTogether": "{{messages}} et {{events}} en attente",
-      "loadingGoal": "Chargement de l'objectif...",
       "message_one": "{{count}} message",
       "message_many": "{{count}} messages",
       "message_other": "{{count}} messages",
-      "openGoalDetails": "Ouvrir les détails de l'objectif",
       "pendingAutomationEvent": "Événement d'automatisation en attente",
       "queuedMessage": "Message en file d'attente",
       "queuedMessageDescription": "Attend en ligne et envoie une fois l'exécution actuelle terminée.",
@@ -2217,8 +2208,6 @@
       "title": "Remplacer le brouillon du compositeur ?"
     },
     "run": {
-      "automatedRun": "Exécution automatisée",
-      "collapseGroupedHistory": "Réduire l'historique des exécutions groupées",
       "collapseWorkHistory": "Réduire l'historique du travail",
       "creditUsage": "Utilisation du crédit",
       "done": {
@@ -2233,15 +2222,9 @@
         "wrappedUp": "Terminé à {{time}}"
       },
       "duration": {
-        "hour_one": "{{count}} heure",
-        "hour_many": "{{count}} heures",
-        "hour_other": "{{count}} heures",
         "hoursShort_one": "{{count}}h",
         "hoursShort_many": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} min",
-        "minute_many": "{{count}} min",
-        "minute_other": "{{count}} min",
         "minutesShort_one": "{{count}}m",
         "minutesShort_many": "{{count}}m",
         "minutesShort_other": "{{count}}m",
@@ -2249,18 +2232,12 @@
         "secondsShort_many": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{duration}} pour {{label}}",
-      "expandGroupedHistory": "Développer l’historique des exécutions groupées",
       "expandWorkHistory": "Développer l'historique du travail",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Mode rapide désactivé",
       "fastModeOn": "Mode rapide activé",
       "fastModeWillBeOff": "Le mode rapide sera désactivé",
       "fastModeWillBeOn": "Le mode rapide sera activé",
-      "goalFor": "Objectif pour {{label}}",
-      "groupedRunsFor_one": "{{count}} exécution pour {{source}}",
-      "groupedRunsFor_many": "{{count}} exécutions pour {{source}}",
-      "groupedRunsFor_other": "{{count}} exécutions pour {{source}}",
       "justNow": "à l'instant",
       "keepGoing": "Continuez",
       "keepGoingAt": "Continuez · {{timestamp}}",
@@ -2286,7 +2263,6 @@
       "viewActivityLogs": "Voir les journaux d'activité",
       "viewLogs": "Voir les journaux d'exécution",
       "waitingIn": "En attente",
-      "worked": "Fonctionné",
       "workedFor": "Fonctionné pendant {{duration}}",
       "workingFor": "En cours depuis {{duration}}"
     },
