+++ turbo/apps/platform/src/i18n/locales/hi-IN/common.json
@@ -2150,26 +2150,17 @@
     "promptDocumentTitle": "प्रॉम्प्ट",
     "queue": {
       "aboutAutomationEvent": "इस ऑटोमेशन इवेंट के बारे में",
-      "aboutGoal": "इस लक्ष्य के बारे में",
       "aboutQueuedMessage": "इस पंक्तिबद्ध संदेश के बारे में",
-      "activeGoal": "सक्रिय लक्ष्य",
       "automationEvent": "ऑटोमेशन घटना",
       "automationEventDescription": "पंक्तिबद्ध संदेशों के पीछे प्रतीक्षा करता है और वर्तमान रन समाप्त होने पर चलता है।",
-      "cancelGoal": "लक्ष्य रद्द करें",
       "cancellationRecoveryPending": "पंक्तिबद्ध कार्य जारी रखने से पहले रद्द किए गए रन को अंतिम रूप दिया जा रहा है।",
       "event_one": "{{count}} घटना",
       "event_other": "{{count}} घटनाएँ",
       "goal": "लक्ष्य",
-      "goalDescription": "कतार ख़त्म होने के बाद चलता है और तब तक चलता रहता है जब तक आप इसे रद्द नहीं कर देते।",
-      "goalLoadFailed": "यह लक्ष्य लोड नहीं किया जा सका",
-      "goalRetry": "संवाद बंद करें और पुनः प्रयास करने के लिए इसे दोबारा खोलें।",
-      "goalUnavailable": "यह लक्ष्य अब उपलब्ध नहीं है.",
       "itemsWaiting": "{{items}} प्रतीक्षा कर रहा है",
       "itemsWaitingTogether": "{{messages}} और {{events}} प्रतीक्षारत हैं",
-      "loadingGoal": "लक्ष्य लोड हो रहा है...",
       "message_one": "{{count}} संदेश",
       "message_other": "{{count}} संदेश",
-      "openGoalDetails": "लक्ष्य विवरण खोलें",
       "pendingAutomationEvent": "लंबित ऑटोमेशन कार्यक्रम",
       "queuedMessage": "पंक्तिबद्ध संदेश",
       "queuedMessageDescription": "लाइन में प्रतीक्षा करता है और वर्तमान रन समाप्त होने पर भेजता है।",
@@ -2181,8 +2172,6 @@
       "title": "कंपोज़र ड्राफ्ट बदलें?"
     },
     "run": {
-      "automatedRun": "स्वचालित रन",
-      "collapseGroupedHistory": "समूहीकृत रन इतिहास को संक्षिप्त करें",
       "collapseWorkHistory": "कार्य इतिहास संक्षिप्त करें",
       "creditUsage": "क्रेडिट उपयोग",
       "done": {
@@ -2197,28 +2186,19 @@
         "wrappedUp": "{{time}} पर समाप्त हुआ"
       },
       "duration": {
-        "hour_one": "{{count}} घंटा",
-        "hour_other": "{{count}} घंटे",
         "hoursShort_one": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} मि",
-        "minute_other": "{{count}} मिनट",
         "minutesShort_one": "{{count}}m",
         "minutesShort_other": "{{count}}m",
         "secondsShort_one": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{label}} के लिए {{duration}}",
-      "expandGroupedHistory": "समूहीकृत रन इतिहास का विस्तार करें",
       "expandWorkHistory": "कार्य इतिहास का विस्तार करें",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "फ़ास्ट मोड बंद",
       "fastModeOn": "फ़ास्ट मोड चालू",
       "fastModeWillBeOff": "फ़ास्ट मोड बंद होगा",
       "fastModeWillBeOn": "फ़ास्ट मोड चालू होगा",
-      "goalFor": "{{label}} के लिए लक्ष्य",
-      "groupedRunsFor_one": "{{count}} रन {{source}} के लिए",
-      "groupedRunsFor_other": "{{count}} रन {{source}} के लिए",
       "justNow": "बस अब",
       "keepGoing": "जारी रखें",
       "keepGoingAt": "चलते रहें · {{timestamp}}",
@@ -2243,7 +2223,6 @@
       "viewActivityLogs": "गतिविधि लॉग देखें",
       "viewLogs": "रन लॉग देखें",
       "waitingIn": "प्रतीक्षा कर रहा है",
-      "worked": "काम किया",
       "workedFor": "{{duration}} के लिए काम किया",
       "workingFor": "{{duration}} से काम जारी"
     },
