+++ turbo/apps/platform/src/i18n/locales/id-ID/common.json
@@ -2147,26 +2147,17 @@
     "promptDocumentTitle": "Prompt",
     "queue": {
       "aboutAutomationEvent": "Tentang peristiwa otomatisasi ini",
-      "aboutGoal": "Tentang tujuan ini",
       "aboutQueuedMessage": "Tentang pesan yang menunggu ini",
-      "activeGoal": "Tujuan aktif",
       "automationEvent": "Peristiwa otomatisasi",
       "automationEventDescription": "Menunggu di belakang pesan yang menunggu dan dijalankan setelah eksekusi saat ini selesai.",
-      "cancelGoal": "Batalkan tujuan",
       "cancellationRecoveryPending": "Menyelesaikan eksekusi yang dibatalkan sebelum melanjutkan pekerjaan dalam antrean.",
       "event_one": "{{count}} peristiwa",
       "event_other": "{{count}} peristiwa",
       "goal": "Tujuan",
-      "goalDescription": "Berjalan setelah antrean kosong dan terus berjalan sampai Anda membatalkannya.",
-      "goalLoadFailed": "Tidak dapat memuat tujuan ini",
-      "goalRetry": "Tutup dialog dan buka lagi untuk mencoba kembali.",
-      "goalUnavailable": "Tujuan ini sudah tidak tersedia.",
       "itemsWaiting": "{{items}} menunggu",
       "itemsWaitingTogether": "{{messages}} dan {{events}} menunggu",
-      "loadingGoal": "Memuat tujuan...",
       "message_one": "{{count}} pesan",
       "message_other": "{{count}} pesan",
-      "openGoalDetails": "Buka detail tujuan",
       "pendingAutomationEvent": "Peristiwa otomatisasi tertunda",
       "queuedMessage": "Pesan dalam antrean",
       "queuedMessageDescription": "Menunggu giliran dan dikirim setelah eksekusi saat ini selesai.",
@@ -2178,8 +2169,6 @@
       "title": "Ganti draf composer?"
     },
     "run": {
-      "automatedRun": "Eksekusi otomatis",
-      "collapseGroupedHistory": "Ciutkan riwayat eksekusi yang dikelompokkan",
       "collapseWorkHistory": "Ciutkan riwayat kerja",
       "creditUsage": "Penggunaan kredit",
       "done": {
@@ -2194,28 +2183,19 @@
         "wrappedUp": "Diselesaikan pada {{time}}"
       },
       "duration": {
-        "hour_one": "{{count}} jam",
-        "hour_other": "{{count}} jam",
         "hoursShort_one": "{{count}}j",
         "hoursShort_other": "{{count}}j",
-        "minute_one": "{{count}} mnt",
-        "minute_other": "{{count}} mnt",
         "minutesShort_one": "{{count}}m",
         "minutesShort_other": "{{count}}m",
         "secondsShort_one": "{{count}}d",
         "secondsShort_other": "{{count}}d"
       },
-      "durationFor": "{{duration}} untuk {{label}}",
-      "expandGroupedHistory": "Kembangkan riwayat eksekusi yang dikelompokkan",
       "expandWorkHistory": "Kembangkan riwayat kerja",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Mode cepat nonaktif",
       "fastModeOn": "Mode cepat aktif",
       "fastModeWillBeOff": "Mode cepat akan nonaktif",
       "fastModeWillBeOn": "Mode cepat akan aktif",
-      "goalFor": "Tujuan untuk {{label}}",
-      "groupedRunsFor_one": "{{count}} eksekusi untuk {{source}}",
-      "groupedRunsFor_other": "{{count}} eksekusi untuk {{source}}",
       "justNow": "baru saja",
       "keepGoing": "Lanjutkan",
       "keepGoingAt": "Lanjutkan · {{timestamp}}",
@@ -2240,7 +2220,6 @@
       "viewActivityLogs": "Lihat log aktivitas",
       "viewLogs": "Lihat log eksekusi",
       "waitingIn": "Menunggu di",
-      "worked": "Dikerjakan",
       "workedFor": "Dikerjakan selama {{duration}}",
       "workingFor": "Bekerja selama {{duration}}"
     },
