+++ turbo/apps/platform/src/i18n/locales/it-IT/common.json
@@ -2184,28 +2184,19 @@
     "promptDocumentTitle": "Prompt",
     "queue": {
       "aboutAutomationEvent": "Informazioni su questo evento di automazione",
-      "aboutGoal": "A proposito di questo obiettivo",
       "aboutQueuedMessage": "Informazioni su questo messaggio in coda",
-      "activeGoal": "Obiettivo attivo",
       "automationEvent": "Evento di automazione",
       "automationEventDescription": "Attende dietro i messaggi in coda ed esegue una volta terminata l'esecuzione corrente.",
-      "cancelGoal": "Annulla obiettivo",
       "cancellationRecoveryPending": "Finalizzazione dell'esecuzione annullata prima di continuare il lavoro in coda.",
       "event_one": "{{count}} evento",
       "event_many": "{{count}} eventi",
       "event_other": "{{count}} eventi",
       "goal": "Obiettivo",
-      "goalDescription": "Viene eseguito dopo che la coda si è esaurita e continua a funzionare finché non la annulli.",
-      "goalLoadFailed": "Impossibile caricare questo obiettivo",
-      "goalRetry": "Chiudi la finestra di dialogo e riaprila per riprovare.",
-      "goalUnavailable": "Questo obiettivo non è più disponibile.",
       "itemsWaiting": "{{items}} in attesa",
       "itemsWaitingTogether": "{{messages}} e {{events}} in attesa",
-      "loadingGoal": "Caricamento obiettivo...",
       "message_one": "{{count}} messaggio",
       "message_many": "{{count}} messaggi",
       "message_other": "{{count}} messaggi",
-      "openGoalDetails": "Apri i dettagli dell'obiettivo",
       "pendingAutomationEvent": "Evento di automazione in sospeso",
       "queuedMessage": "Messaggio in coda",
       "queuedMessageDescription": "Attende in coda e viene inviato al termine dell'esecuzione corrente.",
@@ -2217,8 +2208,6 @@
       "title": "Sostituire la bozza nell'editor?"
     },
     "run": {
-      "automatedRun": "Esecuzione automatizzata",
-      "collapseGroupedHistory": "Comprimi la cronologia delle esecuzioni raggruppate",
       "collapseWorkHistory": "Comprimi la cronologia lavorativa",
       "creditUsage": "Utilizzo del credito",
       "done": {
@@ -2233,15 +2222,9 @@
         "wrappedUp": "Concluso alle {{time}}"
       },
       "duration": {
-        "hour_one": "{{count}} ora",
-        "hour_many": "{{count}} ore",
-        "hour_other": "{{count}} ore",
         "hoursShort_one": "{{count}}h",
         "hoursShort_many": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} min",
-        "minute_many": "{{count}} min",
-        "minute_other": "{{count}} min",
         "minutesShort_one": "{{count}}m",
         "minutesShort_many": "{{count}}m",
         "minutesShort_other": "{{count}}m",
@@ -2249,18 +2232,12 @@
         "secondsShort_many": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{duration}} per {{label}}",
-      "expandGroupedHistory": "Espandi la cronologia delle esecuzioni raggruppate",
       "expandWorkHistory": "Espandi la storia lavorativa",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Modalità veloce disattivata",
       "fastModeOn": "Modalità veloce attiva",
       "fastModeWillBeOff": "La modalità veloce verrà disattivata",
       "fastModeWillBeOn": "La modalità veloce verrà attivata",
-      "goalFor": "Obiettivo per {{label}}",
-      "groupedRunsFor_one": "{{count}} esecuzione per {{source}}",
-      "groupedRunsFor_many": "{{count}} esecuzioni per {{source}}",
-      "groupedRunsFor_other": "{{count}} esecuzioni per {{source}}",
       "justNow": "proprio adesso",
       "keepGoing": "Continua così",
       "keepGoingAt": "Continua · {{timestamp}}",
@@ -2286,7 +2263,6 @@
       "viewActivityLogs": "Visualizza i registri delle attività",
       "viewLogs": "Visualizza i registri di esecuzione",
       "waitingIn": "In attesa",
-      "worked": "Ha funzionato",
       "workedFor": "Ha lavorato per {{duration}}",
       "workingFor": "In lavorazione da {{duration}}"
     },
