+++ turbo/apps/platform/src/i18n/locales/ja-JP/common.json
@@ -2147,26 +2147,17 @@
     "promptDocumentTitle": "プロンプト",
     "queue": {
       "aboutAutomationEvent": "このオートメーションイベントについて",
-      "aboutGoal": "この目標について",
       "aboutQueuedMessage": "このキューに入れられたメッセージについて",
-      "activeGoal": "アクティブな目標",
       "automationEvent": "オートメーションイベント",
       "automationEventDescription": "キューに入れられたメッセージの後ろで待機し、現在の実行が終了すると実行されます。",
-      "cancelGoal": "ゴールをキャンセルする",
       "cancellationRecoveryPending": "キュー内の作業を続行する前に、キャンセルされた実行の終了処理を行っています。",
       "event_one": "{{count}} イベント",
       "event_other": "{{count}} イベント",
       "goal": "目標",
-      "goalDescription": "キューが空になった後に実行され、キャンセルされるまで実行され続けます。",
-      "goalLoadFailed": "この目標を読み込めませんでした",
-      "goalRetry": "ダイアログを閉じて、もう一度開いて再試行してください。",
-      "goalUnavailable": "この目標はもう利用できません。",
       "itemsWaiting": "{{items}} 待機中",
       "itemsWaitingTogether": "{{messages}} と {{events}} が待機中です",
-      "loadingGoal": "目標を読み込んでいます...",
       "message_one": "{{count}} メッセージ",
       "message_other": "{{count}} メッセージ",
-      "openGoalDetails": "目標の詳細を開く",
       "pendingAutomationEvent": "保留中のオートメーションイベント",
       "queuedMessage": "キューに入れられたメッセージ",
       "queuedMessageDescription": "列に並んで待機し、現在の実行が終了したら送信します。",
@@ -2178,8 +2169,6 @@
       "title": "コンポーザーの下書きを置き換えますか？"
     },
     "run": {
-      "automatedRun": "自動実行",
-      "collapseGroupedHistory": "グループ化された実行履歴を折りたたむ",
       "collapseWorkHistory": "作業履歴を折りたたむ",
       "creditUsage": "クレジット使用量",
       "done": {
@@ -2194,28 +2183,19 @@
         "wrappedUp": "{{time}} で終了"
       },
       "duration": {
-        "hour_one": "{{count}} 時間",
-        "hour_other": "{{count}} 時間",
         "hoursShort_one": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} 分",
-        "minute_other": "{{count}} 分",
         "minutesShort_one": "{{count}}m",
         "minutesShort_other": "{{count}}m",
         "secondsShort_one": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{label}} の場合は {{duration}}",
-      "expandGroupedHistory": "グループ化された実行履歴を展開する",
       "expandWorkHistory": "作業履歴を展開",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Fastモード オフ",
       "fastModeOn": "Fastモード オン",
       "fastModeWillBeOff": "Fastモードがオフになります",
       "fastModeWillBeOn": "Fastモードがオンになります",
-      "goalFor": "{{label}} の目標",
-      "groupedRunsFor_one": "{{count}} は {{source}} に対して実行されます",
-      "groupedRunsFor_other": "{{count}} は {{source}} に対して実行されます",
       "justNow": "たった今",
       "keepGoing": "続ける",
       "keepGoingAt": "続ける · {{timestamp}}",
@@ -2240,7 +2220,6 @@
       "viewActivityLogs": "アクティビティログを表示する",
       "viewLogs": "実行ログを表示する",
       "waitingIn": "待機中",
-      "worked": "働いた",
       "workedFor": "{{duration}} で働いていました",
       "workingFor": "{{duration}} 作業中"
     },
