+++ turbo/apps/platform/src/i18n/locales/ko-KR/common.json
@@ -2147,26 +2147,17 @@
     "promptDocumentTitle": "프롬프트",
     "queue": {
       "aboutAutomationEvent": "이 자동화 이벤트에 대해",
-      "aboutGoal": "이 목표에 대해",
       "aboutQueuedMessage": "이 대기 중인 메시지에 대해",
-      "activeGoal": "활성 목표",
       "automationEvent": "자동화 이벤트",
       "automationEventDescription": "대기 중인 메시지 뒤에서 대기하며 현재 실행이 끝나면 한 번 실행됩니다.",
-      "cancelGoal": "목표 취소",
       "cancellationRecoveryPending": "대기 중인 작업을 계속하기 전에 취소된 실행을 마무리하고 있습니다.",
       "event_one": "{{count}}개 이벤트",
       "event_other": "{{count}}개 이벤트",
       "goal": "목표",
-      "goalDescription": "대기열이 비워진 후 실행되며 취소할 때까지 계속 실행됩니다.",
-      "goalLoadFailed": "이 목표를 불러올 수 없습니다",
-      "goalRetry": "대화 상자를 닫았다가 다시 열어 재시도하세요.",
-      "goalUnavailable": "이 목표는 더 이상 사용할 수 없습니다.",
       "itemsWaiting": "{{items}}개 대기 중",
       "itemsWaitingTogether": "{{messages}}개 메시지와 {{events}}개 이벤트 대기 중",
-      "loadingGoal": "목표 불러오는 중...",
       "message_one": "{{count}}개 메시지",
       "message_other": "{{count}}개 메시지",
-      "openGoalDetails": "목표 세부 정보 열기",
       "pendingAutomationEvent": "대기 중인 자동화 이벤트",
       "queuedMessage": "대기 중인 메시지",
       "queuedMessageDescription": "대기열에서 대기하며 현재 실행이 끝나면 전송됩니다.",
@@ -2178,8 +2169,6 @@
       "title": "작성 중인 초안을 교체하시겠습니까?"
     },
     "run": {
-      "automatedRun": "자동 실행",
-      "collapseGroupedHistory": "그룹화된 실행 기록 접기",
       "collapseWorkHistory": "작업 기록 접기",
       "creditUsage": "크레딧 사용량",
       "done": {
@@ -2194,28 +2183,19 @@
         "wrappedUp": "{{time}}에 마무리됨"
       },
       "duration": {
-        "hour_one": "{{count}}시간",
-        "hour_other": "{{count}}시간",
         "hoursShort_one": "{{count}}시간",
         "hoursShort_other": "{{count}}시간",
-        "minute_one": "{{count}}분",
-        "minute_other": "{{count}}분",
         "minutesShort_one": "{{count}}분",
         "minutesShort_other": "{{count}}분",
         "secondsShort_one": "{{count}}초",
         "secondsShort_other": "{{count}}초"
       },
-      "durationFor": "{{label}}에 대해 {{duration}}",
-      "expandGroupedHistory": "그룹화된 실행 기록 펼치기",
       "expandWorkHistory": "작업 기록 펼치기",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Fast 모드 꺼짐",
       "fastModeOn": "Fast 모드 켜짐",
       "fastModeWillBeOff": "Fast 모드가 꺼집니다",
       "fastModeWillBeOn": "Fast 모드가 켜집니다",
-      "goalFor": "{{label}}의 목표",
-      "groupedRunsFor_one": "{{source}}의 {{count}}회 실행",
-      "groupedRunsFor_other": "{{source}}의 {{count}}회 실행",
       "justNow": "방금",
       "keepGoing": "계속 진행",
       "keepGoingAt": "계속 진행 · {{timestamp}}",
@@ -2240,7 +2220,6 @@
       "viewActivityLogs": "활동 로그 보기",
       "viewLogs": "실행 로그 보기",
       "waitingIn": "대기 중",
-      "worked": "작업 완료",
       "workedFor": "{{duration}} 동안 작업 완료",
       "workingFor": "{{duration}} 동안 작업 중"
     },
