+++ turbo/apps/platform/src/i18n/locales/pt-BR/common.json
@@ -2184,28 +2184,19 @@
     "promptDocumentTitle": "Prompt",
     "queue": {
       "aboutAutomationEvent": "Sobre este evento de automação",
-      "aboutGoal": "Sobre este objetivo",
       "aboutQueuedMessage": "Sobre esta mensagem na fila",
-      "activeGoal": "Objetivo ativo",
       "automationEvent": "Evento de automação",
       "automationEventDescription": "Aguarda atrás das mensagens na fila e é executado quando a execução atual termina.",
-      "cancelGoal": "Cancelar objetivo",
       "cancellationRecoveryPending": "Finalizando a execução cancelada antes de continuar o trabalho na fila.",
       "event_one": "{{count}} evento",
       "event_many": "{{count}} eventos",
       "event_other": "{{count}} eventos",
       "goal": "Objetivo",
-      "goalDescription": "Começa depois que a fila esvazia e continua em execução até você cancelar.",
-      "goalLoadFailed": "Não foi possível carregar este objetivo",
-      "goalRetry": "Feche a janela e abra-a novamente para tentar de novo.",
-      "goalUnavailable": "Este objetivo não está mais disponível.",
       "itemsWaiting": "{{items}} aguardando",
       "itemsWaitingTogether": "{{messages}} e {{events}} aguardando",
-      "loadingGoal": "Carregando objetivo...",
       "message_one": "{{count}} mensagem",
       "message_many": "{{count}} mensagens",
       "message_other": "{{count}} mensagens",
-      "openGoalDetails": "Abrir detalhes do objetivo",
       "pendingAutomationEvent": "Evento de automação pendente",
       "queuedMessage": "Mensagem na fila",
       "queuedMessageDescription": "Aguarda na fila e é enviada quando a execução atual termina.",
@@ -2217,8 +2208,6 @@
       "title": "Substituir rascunho do editor?"
     },
     "run": {
-      "automatedRun": "Execução automatizada",
-      "collapseGroupedHistory": "Recolher histórico de execuções agrupadas",
       "collapseWorkHistory": "Recolher histórico de trabalho",
       "creditUsage": "Uso de créditos",
       "done": {
@@ -2233,15 +2222,9 @@
         "wrappedUp": "Concluído em {{time}}"
       },
       "duration": {
-        "hour_one": "{{count}} hora",
-        "hour_many": "{{count}} horas",
-        "hour_other": "{{count}} horas",
         "hoursShort_one": "{{count}}h",
         "hoursShort_many": "{{count}}h",
         "hoursShort_other": "{{count}}h",
-        "minute_one": "{{count}} min",
-        "minute_many": "{{count}} min",
-        "minute_other": "{{count}} min",
         "minutesShort_one": "{{count}}min",
         "minutesShort_many": "{{count}}min",
         "minutesShort_other": "{{count}}min",
@@ -2249,18 +2232,12 @@
         "secondsShort_many": "{{count}}s",
         "secondsShort_other": "{{count}}s"
       },
-      "durationFor": "{{duration}} para {{label}}",
-      "expandGroupedHistory": "Expandir histórico de execuções agrupadas",
       "expandWorkHistory": "Expandir histórico de trabalho",
       "fastModelName": "{{model}} Fast",
       "fastModeOff": "Modo rápido desativado",
       "fastModeOn": "Modo rápido ativado",
       "fastModeWillBeOff": "O modo rápido será desativado",
       "fastModeWillBeOn": "O modo rápido será ativado",
-      "goalFor": "Objetivo para {{label}}",
-      "groupedRunsFor_one": "{{count}} execução para {{source}}",
-      "groupedRunsFor_many": "{{count}} execuções para {{source}}",
-      "groupedRunsFor_other": "{{count}} execuções para {{source}}",
       "justNow": "agora mesmo",
       "keepGoing": "Continuar",
       "keepGoingAt": "Continuar · {{timestamp}}",
@@ -2286,7 +2263,6 @@
       "viewActivityLogs": "Ver logs de atividade",
       "viewLogs": "Ver logs da execução",
       "waitingIn": "Aguardando na",
-      "worked": "Trabalhou",
       "workedFor": "Trabalhou por {{duration}}",
       "workingFor": "Trabalhando há {{duration}}"
     },
