+++ turbo/apps/platform/src/signals/chat-page/chat-conversation-locator.ts
@@ -19,30 +19,20 @@ import {
   type State,
 } from "ccstate";
 import { animationFrame, timeout } from "signal-timers";
-import { FeatureSwitchKey } from "@okouai/core/feature-switch-key";
 import {
-  applyCompletedWorkExpansion,
-  buildCompletedWorkFolding,
   chatEventDisplayError,
-  completedWorkExpandedKeys$,
-  completedWorkExpandedKeysForScrollTarget,
   isRenderableAssistantEvent,
-} from "./completed-work-folding.ts";
+} from "./chat-event-display.ts";
 import {
   applyRunWorkExpansion,
   buildRunWorkFolding,
   runWorkExpandedKeys$,
   runWorkExpandedKeysForScrollTarget,
 } from "./run-work-folding.ts";
-import { featureSwitch$ } from "../external/feature-switch.ts";
 import { logger } from "../log.ts";
 import { messageDocumentToDisplayText } from "../okou-page/user-message-document-codec.ts";
 import { onDomEventFn, onRef, resetSignal } from "../utils.ts";
 import type { ChatEventGroup, EnrichedChatEvent } from "./chat-event.ts";
-import {
-  buildRunGroupFolding,
-  runGroupExpansionOverrides$,
-} from "./run-group-folding.ts";
 import type {
   ScrollToEventOptions,
   ThreadScrollPosition,
@@ -359,46 +349,14 @@ function createVisibleTurns(
   return computed((get): readonly LocatorTurn[] => {
     const activeGroups = activeGroupsForLocator(get(allChatGroups$));
     const targetEventId = get(threadScrollPosition$)?.targetEventId ?? null;
-    const runWorkFoldingEnabled =
-      get(featureSwitch$)[FeatureSwitchKey.ChatRunWorkFolding] ?? false;
-    const runGroupFolding = buildRunGroupFolding(
-      activeGroups,
-      get(runGroupExpansionOverrides$),
-      targetEventId,
-      { preserveRunGroupsForWorkFolding: runWorkFoldingEnabled },
-    );
-    const runGroupVisibleGroups =
-      runGroupFolding?.visibleGroups ?? activeGroups;
-    if (!runWorkFoldingEnabled) {
-      const completedWorkFolding = buildCompletedWorkFolding(
-        runGroupVisibleGroups,
-      );
-      const completedWorkExpandedKeys =
-        completedWorkExpandedKeysForScrollTarget(
-          completedWorkFolding,
-          get(completedWorkExpandedKeys$),
-          targetEventId,
-        );
-      return turnsFromGroups(
-        applyCompletedWorkExpansion(
-          runGroupVisibleGroups,
-          completedWorkFolding,
-          completedWorkExpandedKeys,
-        ),
-      );
-    }
-    const runWorkFolding = buildRunWorkFolding(runGroupVisibleGroups);
+    const runWorkFolding = buildRunWorkFolding(activeGroups);
     const runWorkExpandedKeys = runWorkExpandedKeysForScrollTarget(
       runWorkFolding,
       get(runWorkExpandedKeys$),
       targetEventId,
     );
     return turnsFromGroups(
-      applyRunWorkExpansion(
-        runGroupVisibleGroups,
-        runWorkFolding,
-        runWorkExpandedKeys,
-      ),
+      applyRunWorkExpansion(activeGroups, runWorkFolding, runWorkExpandedKeys),
     );
   });
 }
