+++ turbo/apps/platform/src/signals/chat-page/chat-event-display.ts
@@ -1,0 +1,39 @@
+import {
+  chatEventCompatibilityRole,
+  isChatEventContentTextType,
+} from "@okouai/api-contracts/contracts/chat-events";
+import { hasChatEventBodyContent } from "./chat-event-body-blocks.ts";
+import type { EnrichedChatEvent } from "./chat-event.ts";
+import type { ChatEvent } from "./chat-event-types.ts";
+
+export function chatEventDisplayError(event: ChatEvent): string | undefined {
+  if (
+    event.eventType === "input.rejected" ||
+    event.eventType === "output.error" ||
+    event.eventType === "run.failed" ||
+    event.eventType === "run.cancelled"
+  ) {
+    return event.error;
+  }
+  return undefined;
+}
+
+function chatEventHasAttachments(event: EnrichedChatEvent): boolean {
+  return (
+    "userMessage" in event &&
+    (event.userMessage?.parts.some((part) => {
+      return part.type === "file";
+    }) ??
+      false)
+  );
+}
+
+export function isRenderableAssistantEvent(event: EnrichedChatEvent): boolean {
+  return (
+    chatEventCompatibilityRole(event.eventType) === "assistant" &&
+    ((isChatEventContentTextType(event.eventType) && Boolean(event.content)) ||
+      Boolean(chatEventDisplayError(event)) ||
+      hasChatEventBodyContent(event) ||
+      chatEventHasAttachments(event))
+  );
+}
