+++ turbo/apps/platform/src/views/okou-page/__tests__/chat-capability-feedback.test.tsx
@@ -186,7 +186,6 @@ test("Hide passage actions for a run summary inside an AI reply", async () => {
     path: RUN_PATH,
     featureSwitches: {
       [FeatureSwitchKey.ChatDesktopSelection]: true,
-      [FeatureSwitchKey.ChatRunWorkFolding]: true,
     },
   });
