+++ turbo/apps/platform/src/views/okou-page/__tests__/chat-capability-run-history.test.tsx
@@ -1,5 +1,4 @@
 import { screen, waitFor } from "@testing-library/react";
-import { FeatureSwitchKey } from "@okouai/core/feature-switch-key";
 import { expect, test } from "vitest";
 
 import {
@@ -184,7 +183,6 @@ test("Project all workflow run outputs through one run-group history", async () 
   await setupPage({
     context,
     path: RUN_PATH,
-    featureSwitches: { [FeatureSwitchKey.ChatRunWorkFolding]: true },
   });
 
   await readyChat();
@@ -193,7 +191,6 @@ test("Project all workflow run outputs through one run-group history", async () 
   expect(screen.queryByText("Earlier workflow evidence 2")).toBeNull();
   const main = screen.getByText("Earlier workflow result 2");
   expect(main).toBeVisible();
-  expect(queryButton("Expand grouped run history")).toBeNull();
   expect(queryWorkHistoryToggle("collapsed")).toBeVisible();
   const currentProgress = await screen.findByLabelText(
     "Checking the latest workflow run",
@@ -245,7 +242,6 @@ test("Project all workflow run outputs through one run-group history", async () 
   expect(currentMain).toBeVisible();
   expect(screen.queryByText("Earlier workflow result 2")).toBeNull();
   expect(currentMain.closest('[data-role="assistant"]')).toBe(assistantGroup);
-  expect(queryButton("Expand grouped run history")).toBeNull();
   await expect(findWorkHistoryToggle("collapsed")).resolves.toBeVisible();
 });
 
@@ -305,7 +301,6 @@ test("Keep different run groups as separate assistant responses", async () => {
   await setupPage({
     context,
     path: RUN_PATH,
-    featureSwitches: { [FeatureSwitchKey.ChatRunWorkFolding]: true },
   });
   await readyChat();
 
@@ -387,7 +382,6 @@ test("Keep the prior goal result as main while the next run has no output", asyn
   await setupPage({
     context,
     path: RUN_PATH,
-    featureSwitches: { [FeatureSwitchKey.ChatRunWorkFolding]: true },
   });
 
   await readyChat();
@@ -395,7 +389,6 @@ test("Keep the prior goal result as main while the next run has no output", asyn
     "The earlier launch evidence is complete.",
   );
   expect(priorMain).toBeVisible();
-  expect(queryButton("Expand grouped run history")).toBeNull();
   expect(queryWorkHistoryToggle("collapsed")).toBeNull();
   const thinking = await waitFor(() => {
     const indicator = document.querySelector<HTMLElement>(
@@ -456,7 +449,6 @@ test("Keep the prior goal result as main while the next run has no output", asyn
     answeringAssistant,
   );
   expect(answeringAssistant).toBe(pendingAssistant);
-  expect(queryButton("Expand grouped run history")).toBeNull();
   expect(
     queryAllByRoleFast("link", answeringAssistant).filter((link) => {
       return link.getAttribute("aria-label") === "View agent profile";
@@ -537,7 +529,6 @@ test("Open an archived goal run from a linked event", async () => {
   await setupPage({
     context,
     path: `${RUN_PATH}#event-${LINKED_ARCHIVED_EVENT_ID}`,
-    featureSwitches: { [FeatureSwitchKey.ChatRunWorkFolding]: true },
   });
 
   await readyChat();
@@ -545,5 +536,4 @@ test("Open an archived goal run from a linked event", async () => {
   expect(linkedResult).toBeVisible();
   expect(screen.getByText("Latest launch result")).toBeVisible();
   expect(queryWorkHistoryToggle("collapsed")).toBeNull();
-  expect(queryButton("Collapse grouped run history")).toBeNull();
 });
