+++ turbo/apps/platform/src/views/okou-page/__tests__/chat-navigation-position-restoration.test.tsx
@@ -1,7 +1,6 @@
 import { fireEvent, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { expect, test } from "vitest";
-import { FeatureSwitchKey } from "@okouai/core/feature-switch-key";
 import { chatThreadEventsContract } from "@okouai/api-contracts/contracts/chat-threads";
 
 import { click, queryAllByRoleFast } from "../../../__tests__/page-helper.ts";
@@ -603,53 +602,47 @@ test("Open a conversation at a linked message", async () => {
   expect(buttonNamed("Scroll to bottom")).toBeInTheDocument();
 });
 
-test.each([false, true])(
-  "Expand folded work when opening a linked message with run work folding %s",
-  async (runWorkFolding) => {
-    const linkedEventId = "folded-link-work-3";
-    const linkedText = "Linked work detail";
-    const events = conversationEvents("folded-link", "Linked")
-      .flatMap((event) => {
-        return event.id === "folded-link-assistant-3"
-          ? [{ ...event, id: linkedEventId, content: linkedText }, event]
-          : [event];
-      })
-      .map((event, index) => {
-        return {
-          ...event,
-          seqId: index + 1,
-          runGroupId:
-            event.runId === "folded-link-run-3" ||
-            event.runId === "folded-link-run-4"
-              ? "folded-link-group"
-              : undefined,
-        };
-      });
-    mockChatLifecycleWithoutBrowserSession({
-      threadId: DEEP_LINK_THREAD_ID,
-      threadTitle: "Linked folded conversation",
-      chatEvents: events,
+test("Expand folded work when opening a linked message", async () => {
+  const linkedEventId = "folded-link-work-3";
+  const linkedText = "Linked work detail";
+  const events = conversationEvents("folded-link", "Linked")
+    .flatMap((event) => {
+      return event.id === "folded-link-assistant-3"
+        ? [{ ...event, id: linkedEventId, content: linkedText }, event]
+        : [event];
+    })
+    .map((event, index) => {
+      return {
+        ...event,
+        seqId: index + 1,
+        runGroupId:
+          event.runId === "folded-link-run-3" ||
+          event.runId === "folded-link-run-4"
+            ? "folded-link-group"
+            : undefined,
+      };
     });
+  mockChatLifecycleWithoutBrowserSession({
+    threadId: DEEP_LINK_THREAD_ID,
+    threadTitle: "Linked folded conversation",
+    chatEvents: events,
+  });
 
-    await setupPage({
-      context,
-      host: APP_HOST,
-      path: `/chats/${DEEP_LINK_THREAD_ID}#event-${linkedEventId}`,
-      featureSwitches: {
-        [FeatureSwitchKey.ChatRunWorkFolding]: runWorkFolding,
-      },
-    });
+  await setupPage({
+    context,
+    host: APP_HOST,
+    path: `/chats/${DEEP_LINK_THREAD_ID}#event-${linkedEventId}`,
+  });
 
-    await waitForThreadMessage(DEEP_LINK_THREAD_ID, linkedText);
-    const geometry = installChatScrollGeometry(
-      threadContainer(DEEP_LINK_THREAD_ID),
-    );
-    fireEvent.resize(window);
+  await waitForThreadMessage(DEEP_LINK_THREAD_ID, linkedText);
+  const geometry = installChatScrollGeometry(
+    threadContainer(DEEP_LINK_THREAD_ID),
+  );
+  fireEvent.resize(window);
 
-    await expectReadingPosition(geometry, linkedText, 0);
-    expect(buttonNamed("Scroll to bottom")).toBeInTheDocument();
-  },
-);
+  await expectReadingPosition(geometry, linkedText, 0);
+  expect(buttonNamed("Scroll to bottom")).toBeInTheDocument();
+});
 
 test("Open a conversation safely when a linked message is unavailable", async () => {
   mockChatLifecycleWithoutBrowserSession({
