+++ turbo/apps/platform/src/views/okou-page/__tests__/chat-sharing.test.tsx
@@ -1,6 +1,5 @@
 import { screen, waitFor, within } from "@testing-library/react";
 import { sharedThreadsContract } from "@okouai/api-contracts/contracts/shared-threads";
-import { FeatureSwitchKey } from "@okouai/core/feature-switch-key";
 import { expect, test } from "vitest";
 
 import {
@@ -176,9 +175,6 @@ test("A folded multi-message answer counts as one shared selection", async () =>
     context,
     path: `/chats/${THREAD_ID}`,
     host: "app.okou.ai",
-    featureSwitches: {
-      [FeatureSwitchKey.ChatRunWorkFolding]: true,
-    },
   });
 
   await screen.findByText("Launch answer 3");
