+++ turbo/apps/platform/src/views/okou-page/chat-composer.tsx
@@ -86,7 +86,6 @@ import {
   SlidersHorizontal,
   Square,
   SwatchBook,
-  Target,
   Trash2,
   User,
   UserCheck,
@@ -250,7 +249,6 @@ import {
   userModelPreference$,
 } from "../../signals/external/user-model-preference.ts";
 import {
-  chatRunWorkFoldingEnabled$,
   codexFastModeEnabled$,
   modelPickerMenuEnabled$,
   customConnectorMcpEnabled$,
@@ -442,139 +440,89 @@ function ComposerQueueGlyph() {
   );
 }
 
-// A single strip row — a queued message, automation event, or active goal. All
-// share one layout so they read as the same kind of pending item; only the
-// leading icon distinguishes them. Goals open a modal because their full
-// objective is fetched lazily by thread.
+// Queued messages and automation events share a pending-item layout.
 function ComposerStripRow({
   kind,
   text,
   onRemove,
-  onOpenDetail,
   removeAriaLabel,
   cancellationRecoveryPending,
 }: {
-  kind: "queued" | "automation-event" | "goal";
+  kind: "queued" | "automation-event";
   text: string;
   onRemove?: () => void;
-  onOpenDetail?: () => void;
   removeAriaLabel: string;
   cancellationRecoveryPending?: boolean;
 }) {
   const { t } = useTranslation();
-  const isGoal = kind === "goal";
   const isAutomationEvent = kind === "automation-event";
-  const itemAriaLabel = isGoal
+  const itemAriaLabel = isAutomationEvent
     ? t(($) => {
-        return $.chat.queue.activeGoal;
+        return $.chat.queue.pendingAutomationEvent;
+      })
+    : t(($) => {
+        return $.chat.queue.queuedMessage;
+      });
+  const aboutAriaLabel = isAutomationEvent
+    ? t(($) => {
+        return $.chat.queue.aboutAutomationEvent;
+      })
+    : t(($) => {
+        return $.chat.queue.aboutQueuedMessage;
+      });
+  const itemTitle = isAutomationEvent
+    ? t(($) => {
+        return $.chat.queue.automationEvent;
+      })
+    : t(($) => {
+        return $.chat.queue.queuedMessage;
+      });
+  const itemDescription = cancellationRecoveryPending
+    ? t(($) => {
+        return $.chat.queue.cancellationRecoveryPending;
       })
     : isAutomationEvent
       ? t(($) => {
-          return $.chat.queue.pendingAutomationEvent;
+          return $.chat.queue.automationEventDescription;
         })
       : t(($) => {
-          return $.chat.queue.queuedMessage;
+          return $.chat.queue.queuedMessageDescription;
         });
-  const aboutAriaLabel = isGoal
-    ? t(($) => {
-        return $.chat.queue.aboutGoal;
-      })
-    : isAutomationEvent
-      ? t(($) => {
-          return $.chat.queue.aboutAutomationEvent;
-        })
-      : t(($) => {
-          return $.chat.queue.aboutQueuedMessage;
-        });
-  const itemTitle = isGoal
-    ? t(($) => {
-        return $.chat.queue.goal;
-      })
-    : isAutomationEvent
-      ? t(($) => {
-          return $.chat.queue.automationEvent;
-        })
-      : t(($) => {
-          return $.chat.queue.queuedMessage;
-        });
-  const itemDescription =
-    cancellationRecoveryPending && !isGoal
-      ? t(($) => {
-          return $.chat.queue.cancellationRecoveryPending;
-        })
-      : isGoal
-        ? t(($) => {
-            return $.chat.queue.goalDescription;
-          })
-        : isAutomationEvent
-          ? t(($) => {
-              return $.chat.queue.automationEventDescription;
-            })
-          : t(($) => {
-              return $.chat.queue.queuedMessageDescription;
-            });
   return (
     <div
       role="listitem"
       aria-label={itemAriaLabel}
       className="group flex items-center gap-2 rounded-md pl-2 pr-1 py-1.5 text-sm text-muted-foreground transition-colors hover:bg-state-hover"
     >
-      {isGoal && onOpenDetail ? (
-        // This target spans almost the whole row, so it deliberately paints no
-        // background of its own — the row's hover carries the highlight and a
-        // second, near-coextensive surface would read as a box inside a box.
-        // Its icon sits in the same p-1 slot the other rows' leading button
-        // uses, keeping every row's glyph and text on one column.
-        <button
-          type="button"
-          className="flex min-w-0 flex-1 items-center gap-2 rounded-md text-left transition-colors hover:text-sidebar-foreground focus-visible:text-sidebar-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-          onClick={onOpenDetail}
-          aria-label={t(($) => {
-            return $.chat.queue.openGoalDetails;
-          })}
+      <Popover>
+        <PopoverTrigger asChild>
+          <button
+            type="button"
+            className="shrink-0 rounded-md p-1 text-emerald-800 transition-colors hover:bg-state-selected-hover focus-visible:bg-state-selected-hover focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
+            aria-label={aboutAriaLabel}
+          >
+            {isAutomationEvent ? (
+              <Bolt size={16} aria-hidden="true" />
+            ) : (
+              <ComposerQueueGlyph />
+            )}
+          </button>
+        </PopoverTrigger>
+        <PopoverContent
+          side="top"
+          align="start"
+          className="w-80 rounded-lg p-3"
         >
-          <span className="flex shrink-0 p-1 text-emerald-800">
-            <Target size={16} aria-hidden="true" />
-          </span>
-          <span className="min-w-0 flex-1 truncate py-1">{text}</span>
-        </button>
-      ) : (
-        <>
-          <Popover>
-            <PopoverTrigger asChild>
-              <button
-                type="button"
-                className="shrink-0 rounded-md p-1 text-emerald-800 transition-colors hover:bg-state-selected-hover focus-visible:bg-state-selected-hover focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-                aria-label={aboutAriaLabel}
-              >
-                {isGoal ? (
-                  <Target size={16} aria-hidden="true" />
-                ) : isAutomationEvent ? (
-                  <Bolt size={16} aria-hidden="true" />
-                ) : (
-                  <ComposerQueueGlyph />
-                )}
-              </button>
-            </PopoverTrigger>
-            <PopoverContent
-              side="top"
-              align="start"
-              className="w-80 rounded-lg p-3"
-            >
-              <p className="text-xs font-semibold text-foreground">
-                {itemTitle}
-              </p>
-              <p className="mt-1 text-xs text-muted-foreground">
-                {itemDescription}
-              </p>
-              <div className="mt-2 max-h-48 overflow-y-auto whitespace-pre-wrap break-words rounded-md bg-muted/50 px-2.5 py-2 text-sm text-foreground">
-                {text}
-              </div>
-            </PopoverContent>
-          </Popover>
-          <span className="min-w-0 flex-1 truncate">{text}</span>
-        </>
-      )}
+          <p className="text-xs font-semibold text-foreground">{itemTitle}</p>
+          <p className="mt-1 text-xs text-muted-foreground">
+            {itemDescription}
+          </p>
+          <div className="mt-2 max-h-48 overflow-y-auto whitespace-pre-wrap break-words rounded-md bg-muted/50 px-2.5 py-2 text-sm text-foreground">
+            {text}
+          </div>
+        </PopoverContent>
+      </Popover>
+      <span className="min-w-0 flex-1 truncate">{text}</span>
       <IconTooltipButton
         type="button"
         className="shrink-0 rounded-lg p-1.5 text-muted-foreground/45 transition-colors hover:bg-state-selected-hover hover:text-sidebar-foreground focus-visible:bg-state-selected-hover focus-visible:text-sidebar-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
@@ -617,19 +565,13 @@ function PendingItemsStripHeader({
 
 function PendingItemsStrip({ signals }: { signals: ComposerSignals }) {
   const { t } = useTranslation();
-  const runWorkFoldingEnabled = useGet(chatRunWorkFoldingEnabled$);
   const pendingEvents =
     useLastResolved(signals.queue.pendingEvents$) ??
     ([] satisfies readonly ComposerPendingEvent[]);
   const cancellationRecoveryPending =
     useLastResolved(signals.queue.cancellationRecoveryPending$) ?? false;
-  const activeGoalObjective = useLastResolved(
-    signals.goal.activeGoalObjective$,
-  );
   const removeQueuedMessage = useSet(signals.queue.removeQueuedMessage$);
   const removeAutomationEvent = useSet(signals.queue.removeAutomationEvent$);
-  const cancelActiveGoal = useSet(signals.goal.cancelActiveGoal$);
-  const openActiveGoal = useSet(signals.goal.openActiveGoal$);
   const pageSignal = useGet(pageSignal$);
   const queued = pendingEvents.filter((event) => {
     return event.kind === "message";
@@ -637,10 +579,6 @@ function PendingItemsStrip({ signals }: { signals: ComposerSignals }) {
   const events = pendingEvents.filter((event) => {
     return event.kind === "automation";
   });
-  const activeGoal =
-    !runWorkFoldingEnabled && activeGoalObjective
-      ? { objective: activeGoalObjective }
-      : undefined;
   const count = queued.length + events.length;
   const messageLabel = t(
     ($) => {
@@ -677,21 +615,16 @@ function PendingItemsStrip({ signals }: { signals: ComposerSignals }) {
             items: queued.length > 0 ? messageLabel : eventLabel,
           },
         );
-  if (count === 0 && !activeGoal) {
+  if (count === 0) {
     return withChatScrollLayout(null);
   }
   return withChatScrollLayout(
     <div className="relative z-0 mx-5 -mb-6 overflow-hidden rounded-xl bg-gray-50 dark:bg-gray-100">
-      {count > 0 ? (
-        <PendingItemsStripHeader
-          label={label}
-          cancellationRecoveryPending={cancellationRecoveryPending}
-        />
-      ) : null}
-      <div
-        className="max-h-[200px] overflow-y-auto px-2 pb-7 pt-1"
-        role={count > 0 || activeGoal ? "list" : undefined}
-      >
+      <PendingItemsStripHeader
+        label={label}
+        cancellationRecoveryPending={cancellationRecoveryPending}
+      />
+      <div className="max-h-[200px] overflow-y-auto px-2 pb-7 pt-1" role="list">
         {queued.map((item) => {
           return (
             <ComposerStripRow
@@ -735,22 +668,6 @@ function PendingItemsStrip({ signals }: { signals: ComposerSignals }) {
             />
           );
         })}
-        {/* The active goal sits last — below queued messages and automation events
-            — because it only runs once the queue drains. Like other pending
-            items it can be cancelled from the strip. */}
-        {activeGoal ? (
-          <ComposerStripRow
-            kind="goal"
-            text={activeGoal.objective}
-            onOpenDetail={openActiveGoal}
-            onRemove={() => {
-              detach(cancelActiveGoal(pageSignal), Reason.DomCallback);
-            }}
-            removeAriaLabel={t(($) => {
-              return $.chat.queue.cancelGoal;
-            })}
-          />
-        ) : null}
       </div>
     </div>,
   );
