+++ turbo/apps/platform/src/views/okou-page/chat-message-surface.tsx
@@ -11,19 +11,16 @@ export const CHAT_THREAD_MESSAGE_LIST_CLASS =
 // use 4px. Section spacing must not become the density of a history/list row.
 export const CHAT_THREAD_RESPONSE_STACK_CLASS = "flex min-w-0 flex-col gap-2";
 export const CHAT_THREAD_RESPONSE_COMPACT_STACK_CLASS =
-  "flex min-w-0 flex-col gap-2 group-data-[run-work-folding]/chat:gap-1";
+  "flex min-w-0 flex-col gap-1";
 
-// Wide response lines align with the 36px avatar regardless of work folding.
-// Folding uses the same frame on narrow layouts. Follow-up items keep their
+// Response lines use a 36px frame at every width. Follow-up items keep their
 // own line metrics.
 export const CHAT_THREAD_RESPONSE_LINE_CLASS =
-  "@[900px]:h-auto @[900px]:min-h-9 @[900px]:py-[calc((2.25rem-1lh)/2)] @[900px]:leading-[1.59375rem] group-data-[run-work-folding]/chat:h-auto group-data-[run-work-folding]/chat:min-h-9 group-data-[run-work-folding]/chat:py-[calc((2.25rem-1lh)/2)] group-data-[run-work-folding]/chat:leading-[1.59375rem]";
+  "h-auto min-h-9 py-[calc((2.25rem-1lh)/2)] leading-[1.59375rem]";
 
-// Bare response icons share the 28px action-button rail when the run-work
-// information architecture is active. Their intrinsic width remains the
-// legacy layout when the feature is off.
+// Bare response icons share the 28px action-button rail.
 export const CHAT_THREAD_RESPONSE_LEADING_ICON_CLASS =
-  "inline-flex shrink-0 items-center justify-center group-data-[run-work-folding]/chat:w-7";
+  "inline-flex w-7 shrink-0 items-center justify-center";
 
 // Keep the entry animation, but do not let its duration also animate the
 // responsive margin: that would continue changing layout after resize.
