+++ turbo/apps/platform/src/views/okou-page/chat-thread-page.tsx
@@ -157,16 +157,9 @@ import {
   closeChatConnectorActionConnectDialog$,
 } from "../../signals/chat-page/connector-action-block.ts";
 import {
-  buildCompletedWorkFolding,
   chatEventDisplayError,
-  completedWorkExpandedKeys$,
-  completedWorkExpandedKeysForScrollTarget,
-  completedWorkFoldForGroup,
   isRenderableAssistantEvent,
-  toggleCompletedWorkExpanded$,
-  type CompletedWorkFold,
-  type CompletedWorkFolding,
-} from "../../signals/chat-page/completed-work-folding.ts";
+} from "../../signals/chat-page/chat-event-display.ts";
 import {
   buildRunWorkFolding,
   runWorkExpandedKeys$,
@@ -176,18 +169,6 @@ import {
   type RunWorkFolding,
   type RunWorkSection,
 } from "../../signals/chat-page/run-work-folding.ts";
-import {
-  buildRunGroupFolding,
-  runGroupExpansionOverrides$,
-  toggleRunGroupExpanded$,
-  type RunGroupFold,
-  type RunGroupFolding,
-} from "../../signals/chat-page/run-group-folding.ts";
-import {
-  activeGoalDialogGoal$,
-  activeGoalDialogThreadId$,
-  closeChatThreadGoalDialog$,
-} from "../../signals/chat-page/chat-goal.ts";
 import { ConnectModal } from "./components/settings/add-connection-dialog.tsx";
 import { CustomConnectorConnectDialog } from "./components/settings/custom-connector-connect-dialog.tsx";
 import {
@@ -3037,14 +3018,6 @@ function ChatThreadScrollCommitMarker({
   );
 }
 
-type WaitingThinkingIndicatorMode = "waiting" | "waiting-queued";
-
-function isWaitingThinkingIndicatorMode(
-  mode: ThinkingIndicatorMode,
-): mode is WaitingThinkingIndicatorMode {
-  return mode === "waiting" || mode === "waiting-queued";
-}
-
 function assistantGroupIdForRunWorkIndicator(
   groups: readonly ChatEventGroup[],
   runWorkFolding: RunWorkFolding | null,
@@ -3077,32 +3050,10 @@ function ChatThreadRenderedEventGroups({
   const modelChanges = modelChangesByEventId(renderedActiveGroups);
   const scrollTargetEventId =
     useGet(thread.threadScrollPosition$)?.targetEventId ?? null;
-  const runWorkFoldingEnabled =
-    useGet(featureSwitch$)[FeatureSwitchKey.ChatRunWorkFolding] ?? false;
-  const runGroupExpansionOverrides = useGet(runGroupExpansionOverrides$);
-  const toggleRunGroupExpanded = useSet(toggleRunGroupExpanded$);
-  const runGroupFolding = buildRunGroupFolding(
+  const runWorkFolding = buildRunWorkFolding(
     renderedActiveGroups,
-    runGroupExpansionOverrides,
-    scrollTargetEventId,
-    { preserveRunGroupsForWorkFolding: runWorkFoldingEnabled },
+    new Set(modelChanges.keys()),
   );
-  const runGroupVisibleGroups =
-    runGroupFolding?.visibleGroups ?? renderedActiveGroups;
-  const completedWorkFolding = runWorkFoldingEnabled
-    ? null
-    : buildCompletedWorkFolding(runGroupVisibleGroups);
-  const completedWorkExpandedKeys = useGet(completedWorkExpandedKeys$);
-  const effectiveCompletedWorkExpandedKeys =
-    completedWorkExpandedKeysForScrollTarget(
-      completedWorkFolding,
-      completedWorkExpandedKeys,
-      scrollTargetEventId,
-    );
-  const toggleCompletedWorkExpanded = useSet(toggleCompletedWorkExpanded$);
-  const runWorkFolding = runWorkFoldingEnabled
-    ? buildRunWorkFolding(runGroupVisibleGroups, new Set(modelChanges.keys()))
-    : null;
   const runWorkExpandedKeys = useGet(runWorkExpandedKeys$);
   const effectiveRunWorkExpandedKeys = runWorkExpandedKeysForScrollTarget(
     runWorkFolding,
@@ -3110,9 +3061,7 @@ function ChatThreadRenderedEventGroups({
     scrollTargetEventId,
   );
   const toggleRunWorkExpanded = useSet(toggleRunWorkExpanded$);
-  const visibleGroups = runWorkFoldingEnabled
-    ? (runWorkFolding?.visibleGroups ?? runGroupVisibleGroups)
-    : (completedWorkFolding?.visibleGroups ?? runGroupVisibleGroups);
+  const visibleGroups = runWorkFolding?.visibleGroups ?? renderedActiveGroups;
   const resolvedThinkingIndicatorMode =
     useLastResolved(thread.thinkingIndicatorMode$) ?? null;
   const thinkingIndicatorMode = runWorkFolding?.statusTail?.events.length
@@ -3122,14 +3071,6 @@ function ChatThreadRenderedEventGroups({
     visibleGroups,
     runWorkFolding,
   );
-  const runGroupFoldPlacements = resolveRunGroupFoldPlacements({
-    groups: visibleGroups,
-    runGroupFolding,
-    onToggleRunGroup: toggleRunGroupExpanded,
-    thinkingIndicatorCanHostFold: isWaitingThinkingIndicatorMode(
-      thinkingIndicatorMode,
-    ),
-  });
 
   return withChatScrollLayout(
     <>
@@ -3137,11 +3078,6 @@ function ChatThreadRenderedEventGroups({
         thread={thread}
         groups={visibleGroups}
         modelChanges={modelChanges}
-        runGroupFoldPlacements={runGroupFoldPlacements}
-        runWorkFoldingEnabled={runWorkFoldingEnabled}
-        completedWorkFolding={completedWorkFolding}
-        completedWorkExpandedKeys={effectiveCompletedWorkExpandedKeys}
-        onToggleCompletedWork={toggleCompletedWorkExpanded}
         runWorkFolding={runWorkFolding}
         runWorkExpandedKeys={effectiveRunWorkExpandedKeys}
         onToggleRunWork={toggleRunWorkExpanded}
@@ -3157,7 +3093,6 @@ function ChatThreadRenderedEventGroups({
         mode={
           runIndicatorAssistantGroupId === null ? thinkingIndicatorMode : null
         }
-        runGroupFolds={runGroupFoldPlacements.thinkingIndicatorRunGroupFolds}
       />
     </>,
   );
@@ -3207,18 +3142,13 @@ function ChatThreadEmptyState({ thread }: { thread: ChatPanelSignals }) {
 }
 
 function ChatThreadEventsMain({ thread }: { thread: ChatPanelSignals }) {
-  const runWorkFoldingEnabled =
-    useGet(featureSwitch$)[FeatureSwitchKey.ChatRunWorkFolding] ?? false;
   const renderedGroupsReady =
     useLastResolved(thread.visibleRenderedChatGroupsReady$) ?? false;
   const scrollContentOnRef = useSet(thread.scrollContentOnRef$);
   const sharingPhase = useGet(thread.sharing.phase$);
 
   return withChatScrollLayout(
-    <main
-      data-run-work-folding={runWorkFoldingEnabled || undefined}
-      className={cn(CHAT_THREAD_CONTENT_MAIN_CLASS, "group/chat")}
-    >
+    <main className={CHAT_THREAD_CONTENT_MAIN_CLASS}>
       <div
         ref={scrollContentOnRef}
         data-message-container
@@ -3240,19 +3170,11 @@ function ChatThreadEventsMain({ thread }: { thread: ChatPanelSignals }) {
 function ChatThreadThinkingIndicator({
   thread,
   mode,
-  runGroupFolds,
 }: {
   thread: ChatPanelSignals;
   mode: ThinkingIndicatorMode;
-  runGroupFolds: readonly RunGroupFoldControl[];
 }) {
-  return (
-    <ThinkingIndicator
-      thread={thread}
-      mode={mode}
-      runGroupFolds={runGroupFolds}
-    />
-  );
+  return <ThinkingIndicator thread={thread} mode={mode} />;
 }
 
 function ChatThreadNextRunModelNotice({
@@ -3311,15 +3233,9 @@ function ChatThreadNextRunModelNotice({
 // messages that visually sit right on top of each other.
 function groupRendersContent(
   group: ChatEventGroup,
-  embeddedFolds: readonly RunGroupFoldControl[],
-  completedWorkFold: CompletedWorkFold | null,
   runWorkSection: RunWorkSection | null,
 ): boolean {
-  if (
-    embeddedFolds.length > 0 ||
-    completedWorkFold !== null ||
-    runWorkSection !== null
-  ) {
+  if (runWorkSection !== null) {
     return true;
   }
   if (group.role === "user") {
@@ -3362,11 +3278,6 @@ function ChatThreadEventGroups({
   thread,
   groups,
   modelChanges,
-  runGroupFoldPlacements,
-  runWorkFoldingEnabled,
-  completedWorkFolding,
-  completedWorkExpandedKeys,
-  onToggleCompletedWork,
   runWorkFolding,
   runWorkExpandedKeys,
   onToggleRunWork,
@@ -3376,19 +3287,12 @@ function ChatThreadEventGroups({
   thread: ChatPanelSignals;
   groups: readonly ChatEventGroup[];
   modelChanges: ReadonlyMap<string, RunModelChange>;
-  runGroupFoldPlacements: RunGroupFoldPlacements;
-  runWorkFoldingEnabled: boolean;
-  completedWorkFolding: CompletedWorkFolding | null;
-  completedWorkExpandedKeys: ReadonlySet<string>;
-  onToggleCompletedWork: (key: string) => void;
   runWorkFolding: RunWorkFolding | null;
   runWorkExpandedKeys: ReadonlySet<string>;
   onToggleRunWork: (key: string) => void;
   thinkingIndicatorMode: ThinkingIndicatorMode;
   runIndicatorAssistantGroupId: string | null;
 }) {
-  const { embeddedRunGroupFolds, externalRunGroupFolds } =
-    runGroupFoldPlacements;
   // A run that ends re-forms the groups around it, so the messages the user
   // sent back to back can land in separate groups with nothing rendered in
   // between. Tracking the last group that actually put something on screen
@@ -3398,34 +3302,14 @@ function ChatThreadEventGroups({
   return (
     <>
       {groups.map((group) => {
-        const runGroupFolds =
-          externalRunGroupFolds.get(group.beginEventId) ?? [];
-        const embeddedFolds =
-          embeddedRunGroupFolds.get(group.beginEventId) ?? [];
-        const completedWorkFold = runWorkFoldingEnabled
-          ? null
-          : completedWorkFoldForGroup(completedWorkFolding, group);
-        const runWorkSection = runWorkFoldingEnabled
-          ? runWorkSectionForGroup(runWorkFolding, group)
-          : null;
+        const runWorkSection = runWorkSectionForGroup(runWorkFolding, group);
         const stackFirstOnPrevious =
-          runGroupFolds.length === 0 &&
           previousVisibleGroup !== undefined &&
           previousVisibleGroup.role === "user" &&
           groupHasUserBubble(previousVisibleGroup);
-        if (
-          groupRendersContent(
-            group,
-            embeddedFolds,
-            completedWorkFold,
-            runWorkSection,
-          )
-        ) {
+        if (groupRendersContent(group, runWorkSection)) {
           previousVisibleGroup = group;
         }
-        const completedWorkExpanded =
-          completedWorkFold !== null &&
-          completedWorkExpandedKeys.has(completedWorkFold.key);
         const runIndicatorMode =
           group.beginEventId === runIndicatorAssistantGroupId &&
           thinkingIndicatorMode !== null
@@ -3436,37 +3320,11 @@ function ChatThreadEventGroups({
             key={runWorkSection?.key ?? group.beginEventId}
             className="contents"
           >
-            {runGroupFolds.map((runGroupFold) => {
-              return (
-                <RunGroupFoldRow
-                  key={runGroupFold.fold.key}
-                  control={runGroupFold}
-                />
-              );
-            })}
             <SelectablePagedGroupRow
               group={group}
               thread={thread}
               modelChanges={modelChanges}
               stackFirstOnPrevious={stackFirstOnPrevious}
-              runGroupFolds={embeddedFolds}
-              completedWorkFold={
-                completedWorkFold !== null
-                  ? {
-                      groups: completedWorkFold.labelGroups,
-                      hiddenGroups: completedWorkFold.hiddenGroups,
-                      expanded: completedWorkExpanded,
-                      onToggle: () => {
-                        if (!completedWorkExpanded) {
-                          captureChatWorkHistoryExpanded({
-                            workStatus: "completed",
-                          });
-                        }
-                        onToggleCompletedWork(completedWorkFold.key);
-                      },
-                    }
-                  : undefined
-              }
               runWorkSection={createRunWorkSectionControl(
                 runWorkSection,
                 runWorkExpandedKeys,
@@ -3486,170 +3344,12 @@ function ChatThreadEventGroups({
   );
 }
 
-interface RunGroupFoldControl {
-  fold: RunGroupFold;
-  expanded: boolean;
-  onToggle: () => void;
-}
-
-interface RunGroupFoldPlacements {
-  embeddedRunGroupFolds: Map<string, RunGroupFoldControl[]>;
-  externalRunGroupFolds: Map<string, RunGroupFoldControl[]>;
-  thinkingIndicatorRunGroupFolds: RunGroupFoldControl[];
-}
-
-function resolveRunGroupFoldPlacements({
-  groups,
-  runGroupFolding,
-  onToggleRunGroup,
-  thinkingIndicatorCanHostFold,
-}: {
-  groups: readonly ChatEventGroup[];
-  runGroupFolding: RunGroupFolding | null;
-  onToggleRunGroup: (key: string, expanded: boolean) => void;
-  thinkingIndicatorCanHostFold: boolean;
-}): RunGroupFoldPlacements {
-  const embeddedRunGroupFolds = new Map<string, RunGroupFoldControl[]>();
-  const externalRunGroupFolds = new Map<string, RunGroupFoldControl[]>();
-  const thinkingIndicatorRunGroupFolds: RunGroupFoldControl[] = [];
-
-  if (runGroupFolding === null) {
-    return {
-      embeddedRunGroupFolds,
-      externalRunGroupFolds,
-      thinkingIndicatorRunGroupFolds,
-    };
-  }
-
-  for (const [index, group] of groups.entries()) {
-    const folds = runGroupFolding.foldsByNextGroupId.get(group.beginEventId);
-    if (!folds || folds.length === 0) {
-      continue;
-    }
-
-    for (const fold of folds) {
-      const control: RunGroupFoldControl = {
-        fold,
-        expanded: fold.expanded,
-        onToggle: () => {
-          onToggleRunGroup(fold.key, fold.expanded);
-        },
-      };
-      if (
-        !control.expanded &&
-        isGoalGroupFold(fold) &&
-        thinkingIndicatorCanHostFold &&
-        waitingIndicatorCanHostCollapsedRunGroupFold(groups, index)
-      ) {
-        thinkingIndicatorRunGroupFolds.push(control);
-        continue;
-      }
-      const embeddedGroupId = control.expanded
-        ? undefined
-        : inlineGroupIdForCollapsedRunGroupFold(groups, index);
-      const target = embeddedGroupId
-        ? embeddedRunGroupFolds
-        : externalRunGroupFolds;
-      const targetGroupId = embeddedGroupId ?? group.beginEventId;
-      const existing = target.get(targetGroupId);
-      if (existing) {
-        existing.push(control);
-      } else {
-        target.set(targetGroupId, [control]);
-      }
-    }
-  }
-
-  return {
-    embeddedRunGroupFolds,
-    externalRunGroupFolds,
-    thinkingIndicatorRunGroupFolds,
-  };
-}
-
-function waitingIndicatorCanHostCollapsedRunGroupFold(
-  groups: readonly ChatEventGroup[],
-  index: number,
-): boolean {
-  const group = groups[index];
-  if (!group || group.role !== "user") {
-    return false;
-  }
-  const runId = firstRunIdForEvents(group.events);
-  if (runId === undefined) {
-    return false;
-  }
-  for (const candidate of groups.slice(index + 1)) {
-    const candidateRunId = firstRunIdForEvents(candidate.events);
-    if (candidateRunId !== undefined && candidateRunId !== runId) {
-      return false;
-    }
-    if (
-      candidateRunId === runId &&
-      candidate.role === "assistant" &&
-      candidate.events.some(isRenderableAssistantEvent)
-    ) {
-      return false;
-    }
-  }
-  return true;
-}
-
-function inlineGroupIdForCollapsedRunGroupFold(
-  groups: readonly ChatEventGroup[],
-  index: number,
-): string | undefined {
-  const group = groups[index];
-  if (!group || group.role !== "user") {
-    return undefined;
-  }
-  if (firstRunIdForEvents(group.events) === undefined) {
-    return undefined;
-  }
-  return (
-    assistantGroupIdForCollapsedRunGroupFold(groups, index) ??
-    group.beginEventId
-  );
-}
-
-function assistantGroupIdForCollapsedRunGroupFold(
-  groups: readonly ChatEventGroup[],
-  index: number,
-): string | undefined {
-  const group = groups[index];
-  if (!group || group.role !== "user") {
-    return undefined;
-  }
-  const runId = firstRunIdForEvents(group.events);
-  if (runId === undefined) {
-    return undefined;
-  }
-
-  for (let nextIndex = index + 1; nextIndex < groups.length; nextIndex++) {
-    const candidate = groups[nextIndex]!;
-    const candidateRunId = firstRunIdForEvents(candidate.events);
-    if (candidateRunId !== runId) {
-      return undefined;
-    }
-    if (candidate.role === "assistant") {
-      return candidate.beginEventId;
-    }
-  }
-
-  return undefined;
-}
-
 function firstRunIdForEvents(
   events: readonly EnrichedChatEvent[],
 ): string | undefined {
   return events.find((event) => {
     return event.runId !== undefined;
   })?.runId;
-}
-
-function parseEventTime(value: string): number | null {
-  const timestamp = Date.parse(value);
-  return Number.isNaN(timestamp) ? null : timestamp;
 }
 
 function formatCompactDuration(totalSeconds: number): string {
@@ -3681,39 +3381,6 @@ function formatCompactDuration(totalSeconds: number): string {
     },
     { count: totalHours },
   );
-}
-
-function durationLabelForGroups(
-  groups: readonly ChatEventGroup[],
-): string | null {
-  const timestamps = groups.flatMap((group) => {
-    return group.events.flatMap((event) => {
-      const timestamp = parseEventTime(event.createdAt);
-      return timestamp === null ? [] : [timestamp];
-    });
-  });
-  if (timestamps.length < 2) {
-    return null;
-  }
-  const elapsedSeconds = Math.max(
-    1,
-    Math.round((Math.max(...timestamps) - Math.min(...timestamps)) / 1000),
-  );
-  return formatCompactDuration(elapsedSeconds);
-}
-
-function completedWorkLabel(groups: readonly ChatEventGroup[]): string {
-  const duration = durationLabelForGroups(groups);
-  return duration
-    ? i18n.t(
-        ($) => {
-          return $.chat.run.workedFor;
-        },
-        { duration },
-      )
-    : i18n.t(($) => {
-        return $.chat.run.worked;
-      });
 }
 
 const RUN_SECTION_LABEL_CLASS =
@@ -3802,51 +3469,6 @@ function FoldedModelChangeDivider({ change }: { change: RunModelChange }) {
       label={modelChangeLabel(t, change)}
       labelPosition="right"
     />
-  );
-}
-
-function CompletedWorkFoldRow({
-  groups,
-  expanded,
-  onToggle,
-}: {
-  groups: readonly ChatEventGroup[];
-  expanded: boolean;
-  onToggle: () => void;
-}) {
-  const { t } = useTranslation();
-  const label = completedWorkLabel(groups);
-  return (
-    <div data-chat-completed-work-fold className="-mx-2">
-      <button
-        type="button"
-        aria-expanded={expanded}
-        aria-label={
-          expanded
-            ? t(($) => {
-                return $.chat.run.collapseWorkHistory;
-              })
-            : t(($) => {
-                return $.chat.run.expandWorkHistory;
-              })
-        }
-        onClick={onToggle}
-        className="inline-flex min-h-9 items-center gap-2 rounded-lg px-2 py-1.5 text-muted-foreground transition-colors hover:bg-state-hover group-data-[run-work-folding]/chat:gap-0"
-      >
-        <span className={CHAT_THREAD_RESPONSE_LEADING_ICON_CLASS}>
-          <Hourglass aria-hidden size={14} />
-        </span>
-        <span className="text-[13px]">{label}</span>
-        <ChevronRight
-          aria-hidden
-          size={14}
-          className={cn(
-            "shrink-0 text-muted-foreground/70 transition-transform group-data-[run-work-folding]/chat:ml-2",
-            expanded && "rotate-90",
-          )}
-        />
-      </button>
-    </div>
   );
 }
 
@@ -3954,63 +3576,6 @@ function RunWorkSectionRow({
   );
 }
 
-function normalizedInlineLabel(value: string): string {
-  return value.trim().replace(/\s+/g, " ");
-}
-
-function runGroupFoldEvents(fold: RunGroupFold): EnrichedChatEvent[] {
-  return fold.labelGroups.flatMap((group) => {
-    return group.events;
-  });
-}
-
-function runGroupFoldSourceLabel(fold: RunGroupFold): string {
-  const events = runGroupFoldEvents(fold);
-  const workflowLabel = runGroupFoldWorkflowLabel(fold);
-  if (workflowLabel) {
-    return workflowLabel;
-  }
-  for (const event of events) {
-    if (!isInputChatEvent(event)) {
-      continue;
-    }
-    const content = messageDocumentToDisplayText(event.userMessage);
-    if (content?.trim()) {
-      return normalizedInlineLabel(content);
-    }
-  }
-  return i18n.t(($) => {
-    return $.chat.run.automatedRun;
-  });
-}
-
-function runGroupFoldWorkflowLabel(fold: RunGroupFold): string | null {
-  for (const event of runGroupFoldEvents(fold)) {
-    const part = eventNonContentPart(event);
-    const label =
-      part?.type === "automation"
-        ? part.automationBrief?.trim() || part.workflowName.trim()
-        : null;
-    if (label) {
-      return normalizedInlineLabel(label);
-    }
-  }
-  return null;
-}
-
-function runGroupFoldGoalLabel(fold: RunGroupFold): string {
-  const goalInputEvent = runGroupFoldEvents(fold).find(isGoalUserMessage);
-  const part = goalInputEvent ? eventNonContentPart(goalInputEvent) : undefined;
-  const content = part?.type === "goal" ? part.goalBrief.trim() : null;
-  return content
-    ? normalizedInlineLabel(content)
-    : i18n
-        .t(($) => {
-          return $.chat.queue.goal;
-        })
-        .toLocaleLowerCase(i18n.resolvedLanguage);
-}
-
 function isRejectedGoalUserMessage(event: EnrichedChatEvent): boolean {
   return (
     event.eventType === "input.rejected" &&
@@ -4025,133 +3590,6 @@ function isGoalUserMessage(
     isInputChatEvent(event) &&
     !isRejectedGoalUserMessage(event) &&
     eventNonContentPart(event)?.type === "goal"
-  );
-}
-
-function isGoalGroupFold(fold: RunGroupFold): boolean {
-  return fold.labelGroups.some((group) => {
-    return group.events.some(isGoalUserMessage);
-  });
-}
-
-function verboseDurationLabelForRunGroupFold(
-  fold: RunGroupFold,
-): string | null {
-  const timestamps = fold.labelGroups.flatMap((group) => {
-    return group.events.flatMap((event) => {
-      if (event.runGroupId !== fold.runGroupId) {
-        return [];
-      }
-      const timestamp = parseEventTime(event.createdAt);
-      return timestamp === null ? [] : [timestamp];
-    });
-  });
-  if (timestamps.length < 2) {
-    return null;
-  }
-  const elapsedMinutes = Math.max(
-    1,
-    Math.round((Math.max(...timestamps) - Math.min(...timestamps)) / 60_000),
-  );
-  const hours = Math.floor(elapsedMinutes / 60);
-  const minutes = elapsedMinutes % 60;
-  const parts: string[] = [];
-  if (hours > 0) {
-    parts.push(
-      i18n.t(
-        ($) => {
-          return $.chat.run.duration.hour;
-        },
-        {
-          count: hours,
-        },
-      ),
-    );
-  }
-  if (minutes > 0 || parts.length === 0) {
-    parts.push(
-      i18n.t(
-        ($) => {
-          return $.chat.run.duration.minute;
-        },
-        {
-          count: minutes,
-        },
-      ),
-    );
-  }
-  return parts.join(" ");
-}
-
-function runGroupFoldLabel(fold: RunGroupFold): string {
-  if (isGoalGroupFold(fold)) {
-    const duration = verboseDurationLabelForRunGroupFold(fold);
-    const label = runGroupFoldGoalLabel(fold);
-    return duration
-      ? i18n.t(
-          ($) => {
-            return $.chat.run.durationFor;
-          },
-          { duration, label },
-        )
-      : i18n.t(
-          ($) => {
-            return $.chat.run.goalFor;
-          },
-          { label },
-        );
-  }
-  const sourceLabel = runGroupFoldSourceLabel(fold);
-  return i18n.t(
-    ($) => {
-      return $.chat.run.groupedRunsFor;
-    },
-    {
-      count: fold.hiddenRunCount,
-      source: sourceLabel,
-    },
-  );
-}
-
-function RunGroupFoldRow({ control }: { control: RunGroupFoldControl }) {
-  const { t } = useTranslation();
-  const { fold, expanded, onToggle } = control;
-  const label = runGroupFoldLabel(fold);
-  const isGoal = isGoalGroupFold(fold);
-  const Icon = isGoal ? Target : Package;
-  return (
-    <div data-chat-run-group-fold className="-mx-2">
-      <button
-        type="button"
-        aria-expanded={expanded}
-        aria-label={
-          expanded
-            ? t(($) => {
-                return $.chat.run.collapseGroupedHistory;
-              })
-            : t(($) => {
-                return $.chat.run.expandGroupedHistory;
-              })
-        }
-        onClick={onToggle}
-        className="inline-flex min-h-9 max-w-full items-center gap-2 rounded-lg px-2 py-1.5 text-muted-foreground transition-colors hover:bg-state-hover group-data-[run-work-folding]/chat:gap-0"
-      >
-        <span className={CHAT_THREAD_RESPONSE_LEADING_ICON_CLASS}>
-          <Icon aria-hidden size={14} />
-        </span>
-        <span className="min-w-0 truncate whitespace-nowrap text-[13px]">
-          {label}
-        </span>
-        <ChevronRight
-          aria-hidden
-          size={14}
-          className={cn(
-            "shrink-0 text-muted-foreground/70 transition-transform group-data-[run-work-folding]/chat:ml-2",
-            expanded && "rotate-90",
-          )}
-        />
-      </button>
-    </div>
   );
 }
 
@@ -4535,7 +3973,7 @@ function RecommendedFollowupList({
               "group flex text-left transition-colors",
               showFollowupCards
                 ? "min-h-24 flex-[0_0_min(22rem,calc(100cqw-4rem))] self-stretch snap-center items-start rounded-[var(--okou-card-radius)] border border-border/70 bg-card p-4 shadow-sm hover:bg-card-hover focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring/50"
-                : "min-h-10 w-full items-center gap-2 rounded-lg px-2 py-2 hover:bg-state-hover group-data-[run-work-folding]/chat:min-h-8 group-data-[run-work-folding]/chat:gap-0 group-data-[run-work-folding]/chat:py-1",
+                : "min-h-8 w-full items-center gap-0 rounded-lg px-2 py-1 hover:bg-state-hover",
             )}
             onClick={() => {
               handleSelect(followup, followupIndex);
@@ -4561,7 +3999,7 @@ function RecommendedFollowupList({
             <ArrowUpRight
               size={14}
               className={cn(
-                "shrink-0 text-muted-foreground/60 opacity-0 transition-all group-hover:text-foreground group-hover:opacity-100 group-data-[run-work-folding]/chat:ml-2",
+                "shrink-0 text-muted-foreground/60 opacity-0 transition-all group-hover:text-foreground group-hover:opacity-100 ml-2",
                 showFollowupCards && "hidden",
               )}
             />
@@ -4618,85 +4056,6 @@ function splitQueuedEventsForThinkingIndicator(groups: ChatEventGroup[]): {
   };
 }
 
-// ---------------------------------------------------------------------------
-// Composer wrapper — reads chat signals from thread prop
-// ---------------------------------------------------------------------------
-
-function ActiveGoalObjectiveDialog({ threadId }: { threadId: string }) {
-  const { t } = useTranslation();
-  const dialogThreadId = useGet(activeGoalDialogThreadId$);
-  const goalLoadable = useLoadable(activeGoalDialogGoal$);
-  const closeDialog = useSet(closeChatThreadGoalDialog$);
-  const open = dialogThreadId === threadId;
-  const goal = goalLoadable.state === "hasData" ? goalLoadable.data : undefined;
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={(nextOpen) => {
-        if (!nextOpen) {
-          closeDialog();
-        }
-      }}
-    >
-      <DialogContent
-        className="w-[calc(100vw-2rem)] max-w-2xl gap-5 p-5 sm:p-6"
-        aria-describedby={undefined}
-      >
-        <DialogHeader>
-          <DialogTitle className="text-base">
-            {t(($) => {
-              return $.chat.queue.goal;
-            })}
-          </DialogTitle>
-          <DialogDescription className="leading-6">
-            {t(($) => {
-              return $.chat.queue.goalDescription;
-            })}
-          </DialogDescription>
-        </DialogHeader>
-        <div className="max-h-[min(60vh,520px)] overflow-y-auto rounded-lg bg-muted/40 px-3 py-3 text-sm text-foreground sm:px-4">
-          {goalLoadable.state === "loading" ? (
-            <div className="flex min-h-28 items-center justify-center gap-2 text-muted-foreground">
-              <Loader2 size={16} className="animate-spin" aria-hidden="true" />
-              <span>
-                {t(($) => {
-                  return $.chat.queue.loadingGoal;
-                })}
-              </span>
-            </div>
-          ) : goalLoadable.state === "hasError" ? (
-            <div className="flex min-h-28 flex-col justify-center gap-1 text-muted-foreground">
-              <p className="font-medium text-foreground">
-                {t(($) => {
-                  return $.chat.queue.goalLoadFailed;
-                })}
-              </p>
-              <p className="text-xs">
-                {t(($) => {
-                  return $.chat.queue.goalRetry;
-                })}
-              </p>
-            </div>
-          ) : goal ? (
-            <Markdown
-              source={goal.objective}
-              escapeHtml
-              style={{ fontSize: "inherit", lineHeight: "inherit" }}
-            />
-          ) : (
-            <div className="flex min-h-28 items-center text-muted-foreground">
-              {t(($) => {
-                return $.chat.queue.goalUnavailable;
-              })}
-            </div>
-          )}
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
 function ChatThreadComposer({ thread }: { thread: ChatPanelSignals }) {
   const composerLayoutRef = useSet(thread.composerLayoutOnRef$);
   const standalonePwa = isStandalonePwa();
@@ -4719,7 +4078,6 @@ function ChatThreadComposer({ thread }: { thread: ChatPanelSignals }) {
       >
         <div className="mx-auto max-w-[900px]">
           <ChatComposer signals={thread.composer} />
-          <ActiveGoalObjectiveDialog threadId={thread.threadId} />
           <PersonalClaudeCodeDeviceAuthDialog />
           <PersonalCodexDeviceAuthDialog />
         </div>
@@ -4804,8 +4162,7 @@ function ShimmerText({
     <p
       ref={setRef}
       className={cn(
-        "okou-shimmer-text h-5 min-w-0 flex-1 truncate text-[0.8125rem] leading-5",
-        "group-data-[run-work-folding]/chat:h-auto group-data-[run-work-folding]/chat:leading-[inherit]",
+        "okou-shimmer-text h-auto min-w-0 flex-1 truncate text-[0.8125rem] leading-[inherit]",
         className,
       )}
       aria-label={ariaLabel}
@@ -4921,7 +4278,7 @@ function InlineThinkingRow({
   return (
     <div
       className={cn(
-        "flex items-center gap-2 h-5 group-data-[run-work-folding]/chat:gap-0",
+        "flex items-center gap-0 h-5",
         CHAT_THREAD_RESPONSE_LINE_CLASS,
       )}
     >
@@ -4987,7 +4344,6 @@ function WaitingForAssistantResponse({
   spinnerEnabled,
   thinkingLabel,
   serverThinkingLabel,
-  runGroupFolds,
   inAssistantGroup,
 }: {
   thread: ChatPanelSignals;
@@ -4996,7 +4352,6 @@ function WaitingForAssistantResponse({
   spinnerEnabled: boolean;
   thinkingLabel: string;
   serverThinkingLabel?: ServerThinkingLabel;
-  runGroupFolds: readonly RunGroupFoldControl[];
   inAssistantGroup: boolean;
 }) {
   const thinkingIndicatorProps = isQueued
@@ -5030,9 +4385,6 @@ function WaitingForAssistantResponse({
       <div className={CHAT_THREAD_ASSISTANT_MESSAGE_ROW_CLASS}>
         <AssistantBubbleAvatar thread={thread} />
         <div className="relative flex min-w-0 flex-col gap-2">
-          {runGroupFolds.map((fold) => {
-            return <RunGroupFoldRow key={fold.fold.key} control={fold} />;
-          })}
           <ChatAssistantMessageBody>
             <InlineThinkingRow
               blockStyle={blockStyle}
@@ -5134,12 +4486,10 @@ function equalRecommendedFollowupSources(
 function ThinkingIndicator({
   thread,
   mode,
-  runGroupFolds,
   inAssistantGroup = false,
 }: {
   thread: ChatPanelSignals;
   mode: ThinkingIndicatorMode;
-  runGroupFolds: readonly RunGroupFoldControl[];
   inAssistantGroup?: boolean;
 }) {
   const featureSwitches = useGet(featureSwitch$);
@@ -5208,7 +4558,6 @@ function ThinkingIndicator({
       spinnerEnabled={spinnerEnabled}
       thinkingLabel={thinkingLabel}
       serverThinkingLabel={serverThinkingLabel}
-      runGroupFolds={runGroupFolds}
       inAssistantGroup={inAssistantGroup}
     />
   );
@@ -5846,7 +5195,7 @@ function AssistantErrorFallback({ error }: { error: string }) {
 
   if (isNoModelProvider) {
     return (
-      <div className="flex items-start gap-2 text-foreground group-data-[run-work-folding]/chat:gap-0">
+      <div className="flex items-start gap-0 text-foreground">
         <AssistantErrorLeadingIcon warning />
         <span>
           {t(($) => {
@@ -5880,7 +5229,7 @@ function AssistantErrorFallback({ error }: { error: string }) {
 
   if (isProviderIncompatible) {
     return (
-      <div className="flex items-start gap-2 text-foreground group-data-[run-work-folding]/chat:gap-0">
+      <div className="flex items-start gap-0 text-foreground">
         <AssistantErrorLeadingIcon warning />
         <span>
           {t(($) => {
@@ -5907,7 +5256,7 @@ function AssistantErrorFallback({ error }: { error: string }) {
 
   if (isProviderDeleted) {
     return (
-      <div className="flex items-start gap-2 text-foreground group-data-[run-work-folding]/chat:gap-0">
+      <div className="flex items-start gap-0 text-foreground">
         <AssistantErrorLeadingIcon warning />
         <span>
           {t(($) => {
@@ -5930,7 +5279,7 @@ function AssistantErrorFallback({ error }: { error: string }) {
   }
 
   return (
-    <div className="flex items-start gap-2 text-destructive group-data-[run-work-folding]/chat:gap-0">
+    <div className="flex items-start gap-0 text-destructive">
       <AssistantErrorLeadingIcon />
       <Markdown
         source={error}
@@ -5988,8 +5337,6 @@ function PagedGroupRow({
   thread,
   modelChanges,
   stackFirstOnPrevious = false,
-  runGroupFolds,
-  completedWorkFold,
   runWorkSection,
   runIndicatorMode,
   statusTailEvents,
@@ -5998,8 +5345,6 @@ function PagedGroupRow({
   thread: ChatPanelSignals;
   modelChanges: ReadonlyMap<string, RunModelChange>;
   stackFirstOnPrevious?: boolean;
-  runGroupFolds?: readonly RunGroupFoldControl[];
-  completedWorkFold?: CompletedWorkFoldControl;
   runWorkSection?: RunWorkSectionControl;
   runIndicatorMode?: Exclude<ThinkingIndicatorMode, null>;
   statusTailEvents?: readonly EnrichedChatEvent[];
@@ -6011,7 +5356,6 @@ function PagedGroupRow({
         thread={thread}
         modelChanges={modelChanges}
         stackFirstOnPrevious={stackFirstOnPrevious}
-        runGroupFolds={runGroupFolds}
       />
     );
   }
@@ -6020,8 +5364,6 @@ function PagedGroupRow({
       group={group}
       thread={thread}
       modelChanges={modelChanges}
-      runGroupFolds={runGroupFolds}
-      completedWorkFold={completedWorkFold}
       runWorkSection={runWorkSection}
       runIndicatorMode={runIndicatorMode}
       statusTailEvents={statusTailEvents}
@@ -6075,8 +5417,6 @@ function SelectablePagedGroupRow({
   thread,
   modelChanges,
   stackFirstOnPrevious,
-  runGroupFolds,
-  completedWorkFold,
   runWorkSection,
   runIndicatorMode,
   statusTailEvents,
@@ -6102,8 +5442,6 @@ function SelectablePagedGroupRow({
         thread={thread}
         modelChanges={modelChanges}
         stackFirstOnPrevious={stackFirstOnPrevious}
-        runGroupFolds={runGroupFolds}
-        completedWorkFold={completedWorkFold}
         runWorkSection={runWorkSection}
         runIndicatorMode={runIndicatorMode}
         statusTailEvents={statusTailEvents}
@@ -6149,8 +5487,6 @@ function SelectablePagedGroupRow({
         thread={thread}
         modelChanges={modelChanges}
         stackFirstOnPrevious={stackFirstOnPrevious}
-        runGroupFolds={runGroupFolds}
-        completedWorkFold={completedWorkFold}
         runWorkSection={runWorkSection}
         runIndicatorMode={runIndicatorMode}
         statusTailEvents={statusTailEvents}
@@ -6178,13 +5514,11 @@ function PagedUserGroup({
   thread,
   modelChanges,
   stackFirstOnPrevious = false,
-  runGroupFolds,
 }: {
   group: ChatEventGroup;
   thread: ChatPanelSignals;
   modelChanges: ReadonlyMap<string, RunModelChange>;
   stackFirstOnPrevious?: boolean;
-  runGroupFolds?: readonly RunGroupFoldControl[];
 }) {
   return (
     <>
@@ -6215,9 +5549,6 @@ function PagedUserGroup({
             />
           </div>
         );
-      })}
-      {runGroupFolds?.map((fold) => {
-        return <RunGroupFoldRow key={fold.fold.key} control={fold} />;
       })}
     </>
   );
@@ -7497,13 +6828,6 @@ function PagedUserMessage({
   );
 }
 
-type CompletedWorkFoldControl = {
-  readonly groups: readonly ChatEventGroup[];
-  readonly hiddenGroups: readonly ChatEventGroup[];
-  readonly expanded: boolean;
-  readonly onToggle: () => void;
-};
-
 type RunWorkSectionControl = Omit<RunWorkSection, "key"> & {
   readonly expanded: boolean;
   readonly onToggle: () => void;
@@ -7513,8 +6837,6 @@ type PagedAssistantGroupProps = {
   readonly group: ChatEventGroup;
   readonly thread: ChatPanelSignals;
   readonly modelChanges: ReadonlyMap<string, RunModelChange>;
-  readonly runGroupFolds?: readonly RunGroupFoldControl[];
-  readonly completedWorkFold?: CompletedWorkFoldControl;
   readonly runWorkSection?: RunWorkSectionControl;
   readonly runIndicatorMode?: Exclude<ThinkingIndicatorMode, null>;
   readonly statusTailEvents?: readonly EnrichedChatEvent[];
@@ -7533,10 +6855,6 @@ type PagedAssistantHistoryItem =
 
 type PagedAssistantTimelineItem =
   | PagedAssistantHistoryItem
-  | {
-      readonly kind: "completed-work";
-      readonly control: CompletedWorkFoldControl;
-    }
   | {
       readonly kind: "run-work";
       readonly control: RunWorkSectionControl;
@@ -7575,23 +6893,12 @@ function foldedRunWorkTimelineItems(
 function buildPagedAssistantTimeline({
   group,
   modelChanges,
-  completedWorkFold,
   runWorkSection,
 }: Pick<
   PagedAssistantGroupProps,
-  "group" | "modelChanges" | "completedWorkFold" | "runWorkSection"
+  "group" | "modelChanges" | "runWorkSection"
 >): PagedAssistantTimelineItem[] {
   const items: PagedAssistantTimelineItem[] = [];
-  if (completedWorkFold !== undefined) {
-    items.push({ kind: "completed-work", control: completedWorkFold });
-    if (completedWorkFold.expanded) {
-      items.push(
-        ...completedWorkFold.hiddenGroups.flatMap((hiddenGroup) => {
-          return assistantTimelineItems(hiddenGroup.events);
-        }),
-      );
-    }
-  }
   if (runWorkSection === undefined) {
     items.push(...assistantTimelineItems(group.events));
     return items;
@@ -7647,16 +6954,6 @@ function PagedAssistantTimeline({
     if (item.kind === "model-change") {
       return (
         <FoldedModelChangeDivider key={item.eventId} change={item.change} />
-      );
-    }
-    if (item.kind === "completed-work") {
-      return (
-        <CompletedWorkFoldRow
-          key="completed-work:control"
-          groups={item.control.groups}
-          expanded={item.control.expanded}
-          onToggle={item.control.onToggle}
-        />
       );
     }
     if (item.kind === "run-work") {
@@ -7717,7 +7014,6 @@ function PagedRunWorkAssistantContent({
   group,
   thread,
   modelChanges,
-  completedWorkFold,
   runWorkSection,
   runIndicatorMode,
   statusTailEvents,
@@ -7726,7 +7022,6 @@ function PagedRunWorkAssistantContent({
   | "group"
   | "thread"
   | "modelChanges"
-  | "completedWorkFold"
   | "runWorkSection"
   | "runIndicatorMode"
   | "statusTailEvents"
@@ -7739,7 +7034,6 @@ function PagedRunWorkAssistantContent({
       }),
     },
     modelChanges,
-    completedWorkFold,
     runWorkSection,
   });
   const mainEvent = runWorkSection
@@ -7784,7 +7078,6 @@ function PagedRunWorkAssistantContent({
             <ThinkingIndicator
               thread={thread}
               mode={runIndicatorMode}
-              runGroupFolds={[]}
               inAssistantGroup
             />
           ) : null}
@@ -7798,8 +7091,6 @@ function PagedAssistantGroup({
   group,
   thread,
   modelChanges,
-  runGroupFolds,
-  completedWorkFold,
   runWorkSection,
   runIndicatorMode,
   statusTailEvents,
@@ -7807,17 +7098,7 @@ function PagedAssistantGroup({
   const hasRenderableEvent = group.events.some((event) => {
     return isRenderableAssistantEvent(event);
   });
-  const hasRunGroupFolds = (runGroupFolds?.length ?? 0) > 0;
-  const visibleCompletedWorkFold = hasRunGroupFolds
-    ? undefined
-    : completedWorkFold;
-  const visibleRunWorkSection = hasRunGroupFolds ? undefined : runWorkSection;
-  if (
-    !hasRenderableEvent &&
-    !completedWorkFold &&
-    !runWorkSection &&
-    !hasRunGroupFolds
-  ) {
+  if (!hasRenderableEvent && !runWorkSection) {
     return null;
   }
 
@@ -7830,7 +7111,7 @@ function PagedAssistantGroup({
     .filter(Boolean)
     .join("\n\n");
   const usesRunWorkPresentation =
-    visibleRunWorkSection !== undefined ||
+    runWorkSection !== undefined ||
     runIndicatorMode !== undefined ||
     (statusTailEvents?.length ?? 0) > 0;
 
@@ -7845,16 +7126,12 @@ function PagedAssistantGroup({
       <div className={CHAT_THREAD_ASSISTANT_MESSAGE_ROW_CLASS}>
         <AssistantBubbleAvatar thread={thread} />
         <div className={cn("relative", CHAT_THREAD_RESPONSE_STACK_CLASS)}>
-          {runGroupFolds?.map((fold) => {
-            return <RunGroupFoldRow key={fold.fold.key} control={fold} />;
-          })}
           {usesRunWorkPresentation ? (
             <PagedRunWorkAssistantContent
               group={group}
               thread={thread}
               modelChanges={modelChanges}
-              completedWorkFold={visibleCompletedWorkFold}
-              runWorkSection={visibleRunWorkSection}
+              runWorkSection={runWorkSection}
               runIndicatorMode={runIndicatorMode}
               statusTailEvents={statusTailEvents}
             />
@@ -7863,7 +7140,6 @@ function PagedAssistantGroup({
               items={buildPagedAssistantTimeline({
                 group,
                 modelChanges,
-                completedWorkFold: visibleCompletedWorkFold,
                 runWorkSection: undefined,
               })}
               thread={thread}
