+++ turbo/packages/core/src/feature-switch-key.ts
@@ -51,7 +51,6 @@ export enum FeatureSwitchKey {
   ZapierConnector = "zapierConnector",
   ComputerUseDesktopPlugins = "computerUseDesktopPlugins",
   ChatErrorRecovery = "chatErrorRecovery",
-  ChatRunWorkFolding = "chatRunWorkFolding",
   PrivateArtifacts = "privateArtifacts",
   AgentMessageMath = "agentMessageMath",
   ProgressiveArtifactPreview = "progressiveArtifactPreview",
