+++ turbo/packages/core/src/feature-switch.ts
@@ -347,12 +347,6 @@ const FEATURE_SWITCHES: Record<FeatureSwitchKey, FeatureSwitch> = {
     enabled: false,
     enabledOrgIdHashes: STAFF_ORG_ID_HASHES,
   },
-  [FeatureSwitchKey.ChatRunWorkFolding]: {
-    maintainer: "ethan@okou.ai",
-    description:
-      "Show live elapsed work status and fold prior assistant output during active and completed chat runs.",
-    enabled: true,
-  },
   // Every artifact privacy slice in #32492 uses this same rollout switch.
   [FeatureSwitchKey.PrivateArtifacts]: {
     maintainer: "yuma@okou.ai",
